Constructs the projection direction with fixed tuning parameter for estimating functional in high-dimensional linear regression model

Constructs the projection direction with fixed tuning parameter for estimating functional in high-dimensional linear regression model

Direction_fixedtuning(Xc, loading, mu = NULL)

Arguments

Xc
Design matrix, of dimension nvar(\(n\))xnobs(\(p\))

loading
observation vector in the linear functional, of length \(p\)

mu
Tuning parameter in construction of projection direction

Value

proj

The projection direction

Examples

X = matrix(sample(-2:2,100*400,replace = TRUE),nrow=100,ncol=400)
Direction_fixedtuning(X,loading=c(1,rep(0,399)),mu=2)

#> [1] "fixed mu"
#> [1] 2
#> [1] -1.391365e-21

#> $proj
#>   [1] -8.147375e-23  5.915362e-24 -4.154547e-24  2.083969e-24  3.759370e-24
#>   [6] -2.210013e-24 -1.513756e-23 -2.193007e-23 -6.241734e-25  5.083409e-24
#>  [11]  1.302186e-24 -4.648623e-24 -4.086643e-24  3.172814e-24 -2.242810e-23
#>  [16] -6.627210e-25 -2.990031e-25 -3.624661e-25 -5.093046e-24 -2.487563e-23
#>  [21]  2.486632e-23  4.469951e-24 -5.371778e-25  2.045826e-23  1.583373e-24
#>  [26] -7.094092e-24 -2.127571e-23  3.142600e-25  1.772004e-24  2.248245e-23
#>  [31]  2.988943e-24  2.398348e-23  1.872517e-23  1.165417e-24 -2.237994e-23
#>  [36]  2.933709e-23  2.195028e-23  2.293912e-23 -2.556654e-23  2.147521e-23
#>  [41]  9.586467e-24 -6.132608e-24 -3.874834e-24  9.293664e-24  4.052262e-24
#>  [46]  4.154777e-25  1.343500e-24  5.772743e-24 -4.528877e-24 -1.179174e-23
#>  [51]  1.652800e-23 -2.437671e-23  3.700744e-24 -2.939267e-24  1.208518e-24
#>  [56]  1.053853e-23  1.987230e-23  5.442721e-24 -2.106685e-23  1.159612e-23
#>  [61] -4.235845e-24  1.672143e-23  3.613903e-24 -2.318677e-23  5.436505e-24
#>  [66]  2.342323e-23  2.072996e-24 -4.085176e-24 -6.236144e-24  2.139745e-23
#>  [71] -3.028628e-24  9.279715e-24 -3.338492e-24  1.880106e-23 -2.022568e-23
#>  [76] -8.077645e-25  4.015995e-24  7.338401e-24  7.821959e-26 -4.535599e-24
#>  [81] -5.683293e-24 -2.721504e-23  2.183297e-24  1.288234e-23  6.196710e-25
#>  [86]  4.661034e-24 -2.157287e-23 -3.473596e-24 -1.850912e-24  5.179321e-24
#>  [91] -4.950007e-24  6.922728e-24 -2.441342e-24  2.125616e-23  7.524134e-24
#>  [96]  1.144905e-24  2.416823e-24  1.032197e-24 -2.795411e-23  1.738665e-24
#> [101] -2.232432e-23 -3.793525e-25  5.221936e-24  2.489797e-24  1.930128e-23
#> [106] -1.301617e-24  5.299413e-24  3.466352e-24  6.739397e-24  1.118962e-24
#> [111] -1.491198e-24 -2.891149e-23  8.198974e-25  1.214031e-23  2.200489e-24
#> [116] -6.174454e-24 -5.217643e-24 -9.962740e-25 -1.406935e-23  2.429491e-23
#> [121] -4.047735e-24  2.053956e-23  2.537311e-23 -5.139926e-24  2.553705e-24
#> [126] -6.161522e-25  4.705267e-24  2.403732e-24 -2.082632e-23  2.005584e-23
#> [131]  6.041173e-24  1.730031e-23  1.329556e-23 -2.366344e-23  3.510573e-24
#> [136] -1.939532e-24 -4.774012e-24  2.058839e-24  2.005828e-23 -4.168529e-24
#> [141] -1.601594e-24  1.389830e-24  7.851777e-24  3.406868e-24 -2.292762e-23
#> [146]  1.870545e-23  4.804549e-24 -2.787781e-23  2.699318e-24 -1.112434e-23
#> [151]  6.501470e-24 -1.519073e-23  9.658224e-24  1.025856e-24  5.658016e-24
#> [156] -3.687765e-24  2.464212e-24 -5.261631e-24  4.482915e-24 -2.462033e-23
#> [161]  2.334337e-24 -7.815887e-24 -2.611319e-24 -1.664202e-24 -4.115919e-24
#> [166]  1.813111e-23 -1.590759e-23 -2.444772e-23 -1.823673e-24 -5.096206e-24
#> [171] -1.692734e-23 -2.449048e-24 -3.856509e-24 -2.750077e-23  2.308106e-23
#> [176]  4.284733e-26  2.469276e-25 -8.168741e-24  4.356457e-24 -2.259333e-24
#> [181]  3.542117e-24 -2.211290e-24  2.267643e-24 -9.654140e-24  8.283519e-25
#> [186] -7.394146e-24  1.461327e-24  2.333319e-23 -5.529127e-25 -3.302537e-25
#> [191]  3.311082e-24  3.846176e-24  3.151370e-24  5.702506e-24 -5.595126e-24
#> [196]  8.735151e-25 -8.235282e-24  2.199855e-23  1.751573e-24 -4.798318e-24
#> [201]  2.825064e-24 -3.374547e-24  6.965179e-24  1.838300e-23 -2.139123e-23
#> [206] -2.462075e-23 -2.672136e-23 -2.167758e-23 -3.309255e-24 -6.586042e-25
#> [211]  8.360825e-24 -7.970133e-25 -6.541371e-24 -6.703438e-24 -6.041765e-25
#> [216] -8.067806e-24 -2.173372e-23 -2.030649e-23 -6.936521e-24 -2.768882e-23
#> [221] -2.076394e-24 -1.873271e-23  1.578870e-23  2.590661e-24 -2.656439e-23
#> [226]  6.130025e-25  6.887657e-25  6.381700e-24 -1.443668e-23 -2.209033e-23
#> [231]  1.933697e-24 -1.386530e-25  4.322144e-24  2.129684e-23  1.521427e-24
#> [236]  1.959844e-23  1.074773e-23  2.521032e-24 -1.198677e-24  1.368469e-23
#> [241]  2.269396e-23  4.566759e-24 -6.503579e-24 -6.494062e-25  1.489159e-24
#> [246]  3.748071e-24  4.443464e-26 -2.604956e-23  4.867961e-24  2.349108e-25
#> [251]  2.030797e-24  1.779365e-23  8.284979e-25 -2.858047e-24  1.903686e-23
#> [256] -2.142566e-23  4.088202e-23  2.412166e-23 -9.296061e-26  2.117389e-23
#> [261] -9.349543e-25 -2.254734e-23 -7.207246e-26  2.848739e-23  5.649719e-24
#> [266]  6.785788e-25 -3.044536e-24  5.216795e-24  7.069713e-26  4.964793e-24
#> [271]  5.829101e-24  1.578332e-23  2.286568e-24 -1.789400e-24 -1.366905e-24
#> [276]  2.203392e-23  6.841654e-25  2.312348e-25  6.241588e-24 -1.378295e-24
#> [281]  2.741777e-24 -1.827992e-23  5.188744e-24  2.699331e-23 -2.315450e-23
#> [286] -6.413224e-24 -2.330697e-24  4.310720e-24  2.579878e-23  1.872337e-23
#> [291]  3.742147e-24 -2.091090e-24  2.526846e-23  9.052179e-25  1.809444e-24
#> [296]  1.072936e-24  2.225380e-24  9.125084e-24 -1.995861e-23 -1.823916e-23
#> [301]  4.331233e-24 -2.177908e-24 -2.848770e-24  2.118089e-23 -2.807201e-24
#> [306]  2.124829e-23  2.039339e-24 -5.064196e-25 -3.258030e-24  2.100677e-23
#> [311]  2.361228e-24 -2.849595e-23 -2.274727e-24 -1.985019e-24  2.823862e-24
#> [316]  1.710887e-24 -3.281724e-24  2.946467e-24  4.864493e-24  7.890399e-25
#> [321]  3.326564e-24 -1.332174e-24  3.160833e-24 -5.544271e-24  5.819703e-24
#> [326] -9.855354e-25 -7.233842e-24  3.639238e-24 -2.010193e-23 -2.499931e-23
#> [331] -2.548171e-25  1.039827e-24 -2.525888e-23  4.089990e-24  5.093517e-25
#> [336] -2.359422e-23 -2.982996e-24  3.163153e-24 -1.913359e-24  1.472781e-24
#> [341] -1.620154e-25  1.936486e-23 -3.624743e-24  2.617762e-24 -5.534989e-24
#> [346] -4.107417e-25  2.709710e-24  1.588805e-23 -3.222720e-24  1.626704e-24
#> [351] -9.251176e-24 -5.573422e-24 -1.007382e-24 -1.785469e-23 -2.818195e-23
#> [356]  2.046148e-24 -7.305241e-25  9.745585e-25  2.617539e-24 -2.831970e-24
#> [361] -1.474548e-26  8.003892e-24  1.742456e-23 -4.739416e-23  2.220927e-23
#> [366] -2.344473e-23 -5.426869e-24 -6.560564e-24  1.225084e-23  6.314436e-24
#> [371] -1.374243e-23 -7.667213e-24 -1.945471e-24 -1.859050e-23  3.192932e-24
#> [376]  1.460505e-24 -3.237649e-24  1.403733e-24  3.899901e-24 -3.875065e-24
#> [381] -1.773870e-24 -2.734112e-23 -2.183414e-23 -5.987199e-24 -7.450229e-24
#> [386]  2.469274e-24 -1.201132e-24 -4.258575e-24 -2.171366e-24  3.581830e-24
#> [391] -2.745668e-23  2.358761e-25  8.073640e-24 -3.019930e-24  2.382048e-23
#> [396] -5.853333e-24  2.425384e-23 -1.869851e-23 -6.435579e-27 -6.920478e-24
#>