Constructs the projection direction with fixed tuning parameter toestimate functional in high-dimensional logistic regression

Constructs the projection direction with fixed tuning parameter toestimate functional in high-dimensional logistic regression

Direction_fixedtuning_logistic(Xc, loading, mu = NULL, weight, f_prime)

Arguments

Xc
Design matrix, of dimension nvar(\(n\))xnobs(\(p\))

loading
observation vector in the linear functional, of length \(p\)

mu
Tuning parameter in construction of projection direction

weight
A vector, of length \(n\), of weights used in re-weighting the correction term of the estimator

f_prime
The first derivative of the logit function at the plug-in LASSO estimate of the linear functional

Value

proj

The projection direction

Examples

X = matrix(sample(-2:2,50*400,replace = TRUE),nrow=50,ncol=400)
y = rbinom(50,1,0.5)
p <- ncol(X);
n <- nrow(X);
col.norm <- 1/sqrt((1/n)*diag(t(X)%*%X)+0.0001);
Xnor <- X %*% diag(col.norm);
fit = glmnet::cv.glmnet(Xnor, y, alpha=1,family = "binomial")
htheta <- as.vector(coef(fit, s = "lambda.min"))
support<-(abs(htheta)>0.001)
Xb <- cbind(rep(1,n),Xnor);
Xc <- cbind(rep(1,n),X);
col.norm <- c(1,col.norm);
pp <- (p+1);
xnew = c(1,rep(0,399))
loading=rep(0,pp)
loading[1]=1
loading[-1]=xnew
htheta <- htheta*col.norm;
htheta <- as.vector(htheta)
f_prime <- exp(Xc%*%htheta)/(1+exp(Xc%*%htheta))^2
Direction_fixedtuning_logistic(Xc,loading,mu=2,weight=rep(1,50),f_prime)

#> [1] "fixed mu"
#> [1] 2

#> $proj
#>   [1] -3.465237e-23 -7.654369e-23  1.995429e-23  1.828068e-24 -6.950662e-26
#>   [6]  1.853782e-23  1.759432e-24 -2.022833e-25  1.934115e-23 -2.062419e-23
#>  [11]  1.423764e-25  1.351800e-24 -1.843354e-23  1.928924e-23  6.432359e-25
#>  [16]  3.002815e-25 -5.080988e-25 -8.147672e-26 -2.017421e-23  1.241192e-25
#>  [21] -2.213587e-24  1.350592e-24  1.241327e-24 -1.950021e-23  3.061931e-25
#>  [26]  1.983310e-23 -1.156349e-24  4.469509e-25  1.997935e-23  9.174670e-25
#>  [31] -1.987340e-23  8.992011e-25 -2.150139e-25 -1.952388e-25  1.482736e-24
#>  [36] -2.146851e-23  2.857096e-25 -1.849460e-23  4.181629e-25  5.010848e-25
#>  [41]  4.295484e-26  1.054771e-24 -2.065492e-23 -6.273855e-25  1.749334e-23
#>  [46]  1.961862e-23 -1.970478e-25 -2.030086e-23 -9.061189e-25 -4.805271e-25
#>  [51] -3.391541e-25  4.498758e-25  3.927434e-25  6.489033e-25  1.922226e-23
#>  [56] -2.026093e-24  1.989993e-24 -2.033267e-23 -1.701907e-24  9.963788e-25
#>  [61]  6.491554e-25 -1.990119e-23  1.022252e-24  1.995802e-23 -8.867711e-26
#>  [66]  5.054138e-25  7.186201e-25  1.526875e-24 -9.708347e-25 -1.306821e-24
#>  [71] -4.351255e-25 -1.992470e-23 -8.866235e-25  4.874356e-25 -1.087994e-24
#>  [76] -2.006678e-23 -1.017631e-24  1.238145e-24  1.840935e-23  1.865849e-23
#>  [81]  1.906721e-23  1.782660e-23 -2.006281e-23 -1.597688e-24  4.575971e-25
#>  [86]  1.864385e-23  1.911946e-23  2.454226e-25  1.671492e-24  4.140809e-25
#>  [91] -1.967099e-23 -1.411516e-25  6.710452e-25  4.619277e-25  1.016694e-24
#>  [96]  2.096049e-23 -4.920682e-25 -2.168163e-23 -3.250973e-24  2.025297e-26
#> [101]  2.108063e-23 -2.002285e-23  7.792646e-25  1.399702e-24  2.249782e-25
#> [106]  1.930313e-23  3.662026e-25 -1.843873e-23  2.110404e-25  1.167322e-24
#> [111] -2.308915e-25 -3.726167e-25 -2.029143e-23  2.108011e-23  5.778771e-25
#> [116]  1.942992e-23  1.244160e-24 -3.205825e-25  7.335995e-25  1.140913e-24
#> [121]  1.829276e-23  1.774727e-23 -7.304051e-25 -1.547544e-24 -1.859096e-24
#> [126]  3.678000e-25  1.624378e-24  7.101616e-26 -8.437991e-26  1.304964e-25
#> [131] -5.799644e-25 -1.510358e-24 -8.023056e-25  2.379086e-24 -9.345550e-25
#> [136] -1.762722e-24 -9.717377e-25  1.072583e-24  7.411242e-25  9.997215e-25
#> [141] -2.170217e-23 -1.986579e-23 -6.957787e-25 -5.105090e-25  1.579922e-24
#> [146]  5.222362e-25  4.755275e-25  6.948840e-25 -1.603418e-24  7.882410e-25
#> [151] -7.192014e-25 -2.554276e-25 -1.681334e-24  2.153512e-25  1.182764e-24
#> [156] -1.118001e-24 -7.402589e-25  2.315726e-24 -1.980066e-23  1.182329e-24
#> [161] -2.017222e-23  2.139794e-24  3.077318e-25  4.624871e-25 -2.202611e-23
#> [166]  3.892966e-25 -4.569559e-25 -6.527136e-25  1.003886e-24  9.418017e-25
#> [171]  1.147906e-24 -1.765320e-24 -5.219334e-25 -1.965211e-23 -8.604689e-25
#> [176] -2.125928e-23  2.713192e-25  1.233499e-24 -2.386092e-24  2.017568e-23
#> [181] -8.892851e-25  7.244367e-25 -1.106223e-24  7.150816e-25 -9.203634e-26
#> [186]  1.175083e-25  1.039844e-24 -8.313642e-25  2.034611e-24 -1.044660e-24
#> [191] -1.964670e-23 -1.282044e-24 -9.747922e-25 -1.056949e-24  2.469851e-25
#> [196]  1.807726e-23  1.935239e-23 -1.254767e-24  1.890874e-23 -6.682364e-25
#> [201] -3.725279e-25 -1.189743e-24 -2.567937e-24  8.593581e-25  9.478115e-25
#> [206]  6.537495e-26  7.222030e-25  5.766623e-25  1.331159e-25 -2.023380e-23
#> [211] -1.651634e-24  7.993931e-25  1.259132e-24  5.513971e-25 -5.667871e-25
#> [216] -1.353312e-24 -2.194297e-23 -2.561762e-25  3.460190e-25 -1.861217e-23
#> [221] -2.021567e-23  1.666980e-24 -1.254858e-24  1.520808e-24  4.934100e-25
#> [226] -8.862476e-25 -2.023021e-23  6.966609e-25  1.174510e-24  4.548585e-25
#> [231]  6.611568e-25  1.899436e-23  1.416462e-26  1.770786e-24 -1.286485e-24
#> [236] -8.572252e-25 -2.181818e-25 -7.075128e-25 -3.844831e-25  1.805848e-23
#> [241] -9.762356e-25 -7.904218e-26 -1.389384e-24  3.367740e-25 -6.435862e-25
#> [246] -5.083633e-25  4.306273e-25  5.060483e-25  4.274651e-25  2.524010e-25
#> [251] -1.139347e-24  8.560019e-26  8.052324e-25 -2.137608e-24  1.413076e-24
#> [256] -2.093456e-23 -1.666479e-27  1.900541e-23 -2.586249e-25  8.132882e-25
#> [261]  2.840037e-24 -1.957180e-23  1.946401e-23  2.157248e-25  4.685954e-26
#> [266] -1.783166e-24 -4.105057e-25  1.532915e-24 -2.021561e-23 -1.269374e-24
#> [271]  3.860739e-25 -1.872942e-23 -1.349827e-24  1.948540e-25 -7.183693e-25
#> [276]  1.879130e-23  2.246124e-25  6.768923e-25  1.952375e-23  6.816956e-25
#> [281]  2.021160e-23 -1.781031e-25 -1.680110e-25 -1.222736e-25  2.826264e-25
#> [286] -4.002774e-25 -9.643808e-25 -1.457809e-24 -2.020305e-23 -4.052593e-25
#> [291]  4.820873e-25 -6.541996e-25  8.456236e-25 -4.306247e-25  1.951915e-23
#> [296] -8.714149e-25 -2.071253e-23  1.300963e-24  1.958153e-23 -1.001282e-25
#> [301]  6.444126e-25  6.094286e-26 -9.365826e-25  1.984427e-23  2.752519e-25
#> [306]  9.645913e-25  1.829212e-24 -2.013851e-23  1.499922e-25 -9.840418e-25
#> [311] -6.174681e-25  3.708858e-25  1.978253e-23  3.232400e-25  9.128907e-25
#> [316] -1.993470e-23 -1.391317e-24 -1.074862e-24 -4.668244e-25 -1.435036e-24
#> [321] -2.108984e-23  6.987434e-25 -1.023468e-24  2.651090e-25  7.979279e-25
#> [326]  3.592408e-25 -2.146168e-23  3.394731e-25  2.374691e-25 -3.396587e-25
#> [331]  6.864235e-25 -1.432234e-24  1.191267e-25  2.055893e-23 -1.883895e-23
#> [336] -1.809015e-23 -8.720087e-26  1.198797e-24  1.959995e-23  3.273053e-25
#> [341] -2.960341e-25  2.210924e-24  1.585879e-24 -2.102596e-23 -1.996624e-23
#> [346]  7.836197e-25 -1.158129e-24  4.078383e-25  1.024370e-24  1.151231e-24
#> [351] -1.652301e-24  1.865150e-23  1.876055e-24  1.994911e-23  1.635277e-24
#> [356]  1.898717e-25  5.807789e-25  3.761208e-25  7.233319e-25  8.092030e-25
#> [361]  1.349639e-24  1.092673e-24  9.682025e-25 -1.899341e-23 -3.238928e-25
#> [366]  8.082121e-25 -6.039101e-25  1.786142e-23  9.009138e-25 -2.442062e-24
#> [371]  1.765941e-23 -1.083655e-24 -4.834641e-25  3.016061e-24  1.865683e-23
#> [376] -1.161134e-24 -1.573620e-24  4.863678e-25  1.969388e-23 -7.162548e-25
#> [381] -6.811931e-25  7.339088e-25  8.485629e-25 -1.996630e-23 -1.952155e-23
#> [386]  1.587969e-24  1.950133e-25  2.240795e-25  3.130140e-25  1.004906e-24
#> [391] -8.161020e-25 -1.086365e-24  1.770954e-23 -2.037789e-24  1.342740e-25
#> [396] -9.856866e-25 -2.134871e-23  3.604362e-25  2.095806e-23 -2.776405e-25
#> [401]  5.000909e-25
#>