Searches for the best step size and computes the projection direction in high dimensional linear regression

Searches for the best step size and computes the projection direction in high dimensional linear regression

Direction_searchtuning(
  X,
  loading,
  model = "linear",
  mu = NULL,
  weight = NULL,
  deriv.vec = NULL,
  resol = 1.5,
  maxiter = 10
)

Arguments

X
Design matrix, of dimension \(n\) x \(p\)

loading
Loading, of length \(p\)

model
The high dimensional regression model, either linear or logistic

mu
The tuning parameter used in the construction of the projection direction

weight
The weight vector of length \(n\), used in correcting the plug-in estimator ; to be supplied if model="logistic" (default=NULL)

deriv.vec
The first derivative vector of the logit function at \(X\widehat{\beta}\) ; to be supplied if model="logistic"

resol
Resolution or the factor by which mu is increased/decreased to obtain the smallest mu that gives convergence of the optimization problem for constructing the projection direction (default = 1.5)

maxiter
maxiter Maximum number of steps along which mu is increased/decreased to obtain the smallest mu that gives convergence of the optimization problem for constructing the projection direction (default = 10)

Value

proj

The projection direction

Examples

n = 100
p = 400
X = matrix(sample(-2:2,n*p,replace = TRUE),nrow = n,ncol = p)
Direction_searchtuning(X,loading=c(1,rep(0,(p-1))))

#> [1] 3

#> $proj
#>   [1]  7.473343e-01 -1.926469e-21  1.724226e-21  3.136533e-22  2.376547e-21
#>   [6]  1.686000e-21 -1.756145e-21  7.525316e-22  2.261419e-22  8.039184e-22
#>  [11] -2.801549e-21 -1.112624e-21 -8.144847e-22 -1.666780e-21  8.332623e-22
#>  [16] -4.406893e-22 -2.516951e-21 -4.712721e-22 -1.752492e-21 -1.044089e-21
#>  [21] -5.664355e-22  1.056222e-21 -1.127764e-21  2.039965e-21  3.201165e-21
#>  [26]  2.059964e-23  1.966146e-21 -1.913273e-21  9.809292e-22  8.867234e-23
#>  [31]  1.116282e-21 -1.475742e-21 -1.829379e-21  3.228311e-22 -2.847748e-21
#>  [36] -2.731973e-21  2.756482e-22  7.872908e-22  1.304888e-22 -9.909675e-24
#>  [41] -4.587682e-23 -2.714577e-21  6.319867e-22 -2.804978e-21 -6.468343e-22
#>  [46]  1.224000e-21 -8.781598e-23 -8.636533e-22 -1.363250e-02  1.227493e-21
#>  [51] -1.435399e-21  1.699471e-21 -1.690966e-21  1.354773e-21  1.926111e-02
#>  [56]  5.681013e-22 -1.313174e-21 -1.854478e-21  1.351819e-21  1.589949e-21
#>  [61] -3.347619e-02 -1.725703e-21  6.216338e-22  8.500242e-22  1.958886e-21
#>  [66] -1.627091e-22  5.284185e-02  3.875736e-21  1.900237e-22 -2.150036e-21
#>  [71]  4.116342e-22 -3.154398e-21  1.584549e-21 -2.084070e-21  2.726935e-21
#>  [76]  1.194772e-21 -1.365804e-02  1.613919e-21  4.978019e-02 -1.283337e-21
#>  [81] -6.461482e-22  2.272001e-21 -1.151250e-21  8.758907e-22 -5.485713e-22
#>  [86] -4.488270e-22 -5.613384e-22  2.839372e-21 -8.209653e-22  5.446024e-22
#>  [91]  5.029423e-23 -2.599442e-21  1.073280e-21  8.481070e-22 -1.206295e-21
#>  [96]  3.187089e-21 -9.580696e-02 -3.553700e-03  5.776933e-21 -1.219263e-21
#> [101] -1.275515e-21  7.318420e-22  3.208189e-21 -3.456304e-21 -2.906145e-22
#> [106]  1.388206e-21 -3.853981e-22 -5.091803e-03 -7.295329e-02  3.224370e-22
#> [111]  2.838234e-21  2.509688e-21 -3.075633e-21 -3.326542e-21  1.375310e-21
#> [116] -5.423116e-22  2.199208e-21 -6.165349e-22 -1.209501e-21 -5.969206e-22
#> [121]  2.739837e-21  1.003356e-21 -1.728484e-22  1.115307e-21 -1.095461e-21
#> [126] -1.685693e-21 -1.306419e-21 -3.940120e-23  8.517053e-22 -2.070198e-21
#> [131] -1.678917e-02 -2.002187e-21 -1.500391e-21  1.244199e-02  2.751147e-21
#> [136]  1.003029e-21  9.518693e-22  1.744536e-21 -8.722173e-22 -5.891082e-22
#> [141] -1.822420e-21 -4.363365e-21  7.128020e-22 -2.099697e-21 -2.556588e-22
#> [146] -3.498985e-21  2.429872e-22 -9.695204e-22  2.130011e-21 -2.963407e-21
#> [151] -1.423737e-21 -4.829174e-03  2.588890e-21  8.785678e-23  6.785271e-22
#> [156]  1.609224e-21  3.169876e-21  2.830735e-21  1.729496e-21 -7.893428e-04
#> [161]  8.549991e-22  1.282684e-21  2.377648e-21 -1.036880e-21  7.844228e-22
#> [166] -1.362759e-21 -2.439271e-21  2.670463e-02 -1.486054e-21 -5.481182e-03
#> [171]  6.703634e-02 -1.015311e-21 -1.499011e-21  1.218894e-21 -1.893815e-22
#> [176]  2.561886e-21  3.799084e-21 -7.000245e-22 -1.058383e-21  2.424833e-02
#> [181]  2.389211e-21  1.827548e-21  7.691680e-22 -2.664311e-21  4.476375e-22
#> [186]  5.518788e-22  5.947654e-02  2.488337e-21  1.050252e-21  3.670047e-22
#> [191] -3.798664e-21 -1.236436e-02  2.554532e-02 -2.579092e-21 -3.305638e-22
#> [196] -4.089245e-03  8.879423e-22  1.911698e-21  1.518706e-21  8.406996e-22
#> [201] -2.400317e-22 -3.333703e-22 -2.347830e-21  1.276541e-21 -2.212350e-21
#> [206]  1.917636e-21 -4.122682e-21  2.640722e-21 -4.117795e-22  3.311359e-03
#> [211]  8.965063e-22 -5.452771e-22  2.134618e-21 -4.466404e-22 -6.789650e-22
#> [216] -2.952811e-21  1.707899e-21 -1.302242e-21  1.238914e-22  8.716639e-22
#> [221]  1.331134e-22 -1.783842e-21 -1.136737e-21  7.535827e-23  2.059793e-21
#> [226]  3.096944e-03 -4.209804e-21  2.104141e-22  9.567499e-22  4.620352e-22
#> [231] -5.584195e-22  8.994189e-22  1.388042e-21  3.892721e-21 -9.898701e-23
#> [236]  1.205256e-21  7.725481e-22 -1.637416e-22  2.644712e-21  1.734285e-21
#> [241]  2.695065e-21  1.884594e-22  4.313624e-21 -1.580076e-21  2.871243e-22
#> [246]  4.576808e-22  1.172886e-21 -1.262887e-21 -8.732012e-22 -2.028424e-21
#> [251] -1.172040e-22 -1.578652e-21 -2.637247e-23  2.054076e-02  9.325474e-22
#> [256] -3.081255e-21 -2.274005e-21 -2.810747e-22 -1.285714e-21 -3.568635e-22
#> [261]  1.038372e-21  6.138244e-02  1.970392e-21  1.908145e-02  9.168303e-25
#> [266] -3.200199e-21 -1.096276e-21 -1.912099e-21 -1.165099e-21 -4.721687e-22
#> [271] -7.978822e-22 -1.468168e-21  1.331813e-21 -2.421025e-21  2.772941e-22
#> [276]  2.443991e-21  6.980858e-02 -3.773659e-21 -1.357878e-21 -1.706791e-22
#> [281] -1.138647e-22  3.876517e-21 -2.043464e-21  9.056254e-03  1.000515e-21
#> [286] -3.327479e-21 -7.119450e-23 -3.581052e-21  2.088844e-22 -1.544903e-21
#> [291] -1.394494e-21 -1.592620e-21  1.749747e-02 -9.976938e-02  1.289395e-21
#> [296] -7.047871e-22  2.578961e-22 -2.438789e-21 -2.630961e-21  1.502785e-21
#> [301]  1.048956e-21  2.093032e-21  3.331009e-21 -4.274488e-21 -3.563237e-21
#> [306] -4.311678e-21  2.257971e-21 -7.491064e-22 -2.188184e-21  1.949478e-02
#> [311] -1.537930e-21  5.024532e-22  2.390078e-21 -8.165161e-22  2.493014e-21
#> [316]  3.103221e-21 -6.578571e-02 -1.644489e-21 -1.735946e-21 -1.033243e-01
#> [321] -2.379786e-21 -2.541974e-22  5.239502e-22 -4.188726e-21  5.769153e-22
#> [326] -9.247989e-22 -9.617784e-22  3.580250e-22  1.769638e-21 -1.978303e-22
#> [331]  1.135224e-21  1.759668e-21 -6.213939e-22  1.251619e-21  1.287036e-21
#> [336]  1.903116e-21 -1.944294e-21 -1.789872e-21  1.411830e-21 -1.836877e-22
#> [341] -2.756991e-21 -1.186470e-21 -1.971370e-21  3.124646e-21 -4.079171e-23
#> [346]  2.571332e-21  5.580251e-22  5.432345e-02 -1.230783e-21 -3.170250e-22
#> [351]  2.262313e-21 -4.351724e-22 -2.259867e-21 -1.069334e-02  1.115249e-21
#> [356]  1.321204e-21  4.285589e-21 -2.756507e-02 -1.396743e-21 -1.091000e-21
#> [361] -7.392077e-22 -2.561334e-21 -1.025036e-03 -1.438591e-21 -1.032860e-21
#> [366] -4.004379e-22 -1.960901e-21 -2.496743e-21  5.525653e-22  5.210736e-21
#> [371]  1.393586e-03  7.980982e-22 -1.196732e-21 -7.272399e-22  2.553976e-21
#> [376]  1.096982e-21 -2.373079e-21  6.169633e-02 -1.249038e-02  2.275467e-21
#> [381]  1.635020e-02  1.409207e-21 -5.678673e-22  1.057267e-21  2.464307e-21
#> [386] -1.595899e-22  3.853550e-22 -6.837535e-22  1.030726e-21  1.167234e-21
#> [391] -1.015716e-21  3.899558e-21  4.353115e-21 -4.481019e-02  7.874890e-22
#> [396] -2.571982e-21 -3.538895e-21  2.971402e-22  5.637951e-22 -3.233559e-22
#> 
#> $step
#> [1] 3
#>