Constructs the projection direction with "optimal" tuning parameter for estimating functional in high-dimensional logistic regression

Constructs the projection direction with "optimal" tuning parameter for estimating functional in high-dimensional logistic regression

Direction_searchtuning_logistic(
  Xc,
  loading,
  mu = NULL,
  weight,
  f_prime,
  resol,
  maxiter
)

Arguments

Xc
Design matrix

loading
observation vector in linear functional

mu
Tuning parameter in construction of the projection direction

weight
A vector of weights used in re-weighting the correction term of the estimator

f_prime
The first derivative of the logit function at the plug-in LASSO estimate of the linear functional

resol
Resolution or the factor by which mu is increased/decreased to obtain the smallest mu that gives convergence of the optimization problem for constructing the projection direction

maxiter
Maximum number of steps along which mu is increased/decreased to obtain the smallest mu that gives convergence of the optimization problem for constructing the projection direction

Value

proj

The projection direction

Examples

X = matrix(sample(-2:2,50*400,replace = TRUE),nrow=50,ncol=400)
y = rbinom(50,1,0.5)
p <- ncol(X);
n <- nrow(X);
col.norm <- 1/sqrt((1/n)*diag(t(X)%*%X)+0.0001);
Xnor <- X %*% diag(col.norm);
fit = glmnet::cv.glmnet(Xnor, y, alpha=1,family = "binomial")
htheta <- as.vector(coef(fit, s = "lambda.min"))
support<-(abs(htheta)>0.001)
Xb <- cbind(rep(1,n),Xnor);
Xc <- cbind(rep(1,n),X);
col.norm <- c(1,col.norm);
pp <- (p+1);
xnew = c(1,rep(0,399))
loading=rep(0,pp)
loading[1]=1
loading[-1]=xnew
htheta <- htheta*col.norm;
htheta <- as.vector(htheta)
f_prime <- exp(Xc%*%htheta)/(1+exp(Xc%*%htheta))^2
Direction_searchtuning_logistic(Xc,loading,weight=rep(1,50),f_prime = f_prime,
                                resol = 1.5, maxiter = 6)

#> [1] 2

#> $proj
#>   [1]  1.627273e+00  1.202323e+00  2.859808e-20 -1.149474e-20  1.255331e-21
#>   [6] -7.226174e-21 -1.876453e-23 -1.876669e-20 -3.370405e-20 -7.173685e-21
#>  [11]  8.540005e-21 -6.965923e-21 -7.328374e-21  2.273468e-20 -1.261258e-20
#>  [16] -3.150713e-21  3.804722e-21 -1.198958e-20 -1.673278e-20  2.729713e-20
#>  [21] -1.024828e-20  1.024679e-21 -3.656201e-20 -7.800050e-21  6.721306e-21
#>  [26]  1.172879e-20  4.476563e-21 -2.296007e-21 -1.637992e-20 -3.447250e-21
#>  [31]  2.882097e-20 -6.246547e-21 -5.370832e-22 -2.731514e-20  1.137300e-20
#>  [36] -1.997217e-21  1.139776e-20 -5.611327e-21  1.072265e-20 -1.841601e-20
#>  [41] -2.244422e-20 -1.084177e-21 -4.577970e-21  7.523291e-21  9.707987e-21
#>  [46] -1.483729e-20  6.985409e-21 -2.480487e-20  4.572885e-21 -2.636585e-20
#>  [51] -7.050576e-21  1.204049e-20  1.258676e-22  1.717287e-20  1.961743e-20
#>  [56] -1.001366e-20 -1.758737e-20  6.865318e-21 -1.858149e-20 -5.095386e-22
#>  [61]  1.407240e-20  1.638881e-20  1.280506e-20  4.396013e-21  1.891987e-20
#>  [66] -2.596101e-21 -7.577362e-21  6.492358e-21 -2.254129e-20  5.354646e-21
#>  [71]  2.076436e-21 -3.981819e-21  2.995630e-21  2.365581e-21  4.341836e-21
#>  [76]  9.971651e-21 -5.090745e-21  1.041846e-20 -5.898556e-21 -5.694052e-21
#>  [81]  3.048555e-20  4.156922e-21 -1.055826e-20 -2.140736e-20 -2.617960e-20
#>  [86] -2.225169e-20  1.874641e-20 -2.474537e-20  1.852613e-20 -2.150684e-21
#>  [91]  6.666805e-21  1.046304e-20  2.003152e-20  3.356620e-21 -4.472351e-21
#>  [96]  2.411346e-21 -1.669779e-20 -3.667178e-22 -1.395966e-20 -1.153888e-20
#> [101]  1.438932e-20 -2.567331e-20  1.526843e-20  3.234061e-21 -1.129080e-21
#> [106]  7.349932e-21 -1.703181e-20 -5.631833e-22  1.334580e-21  1.163110e-20
#> [111]  6.992018e-21  5.820479e-21 -1.762921e-20 -1.686117e-20 -1.961329e-20
#> [116] -8.447976e-21  2.945458e-21 -9.352745e-21  1.814849e-21  1.729436e-20
#> [121]  9.402016e-21 -2.720263e-20  5.150435e-21  1.919310e-20 -2.040480e-21
#> [126]  3.761956e-20  4.059797e-21  1.916442e-20 -3.953114e-20  1.006483e-20
#> [131] -1.876812e-20 -2.733698e-21  8.401859e-21 -1.798474e-21 -6.600720e-22
#> [136]  3.059561e-20 -1.170277e-20  2.230476e-20 -2.169864e-20  1.579501e-20
#> [141] -2.830756e-20  1.576960e-01 -1.273619e-20 -4.590990e-21 -1.148776e-20
#> [146]  2.986959e-21  1.481930e-20  2.604021e-20 -2.549617e-20 -1.755026e-20
#> [151]  3.487540e-20 -1.150011e-20  1.533658e-20  5.671256e-21 -1.195368e-20
#> [156] -6.888037e-21 -3.204618e-21 -6.258867e-22 -1.290733e-20  4.575213e-21
#> [161] -1.099601e-21  2.422841e-20  1.340754e-20 -1.828787e-20  2.177255e-20
#> [166]  5.397088e-21 -1.120007e-21  2.541734e-20  2.441128e-20 -1.373622e-20
#> [171]  1.940500e-21 -1.915051e-20 -2.779246e-20  9.376209e-21  6.494374e-21
#> [176]  2.688216e-20  1.361900e-21 -1.878328e-20 -2.453114e-20  5.320250e-21
#> [181] -1.768081e-21  4.126242e-21  2.100636e-21 -1.016866e-20 -1.607895e-20
#> [186]  1.546855e-20  2.691482e-20 -3.108048e-20 -2.119684e-20  4.832782e-21
#> [191] -2.436754e-21 -1.559886e-20 -6.015396e-21  1.649423e-20  9.949781e-21
#> [196]  1.188455e-20  3.236017e-21 -8.908162e-21  5.736808e-21 -5.777994e-21
#> [201]  7.599280e-21 -1.129237e-20  3.198882e-21  2.566148e-20  1.598654e-20
#> [206] -1.209260e-22 -1.824292e-20 -1.087951e-20 -8.528088e-21 -5.932171e-21
#> [211]  3.270680e-21  1.395837e-20  1.871033e-20  2.243116e-20 -3.506792e-21
#> [216]  2.001894e-21  6.658695e-21  1.962961e-20  5.773447e-21  1.319189e-20
#> [221]  5.273189e-21 -1.625927e-20  8.120420e-22  1.634008e-20 -2.220197e-20
#> [226]  2.402571e-20  3.410639e-21  1.549458e-21 -1.116682e-20  1.571025e-20
#> [231]  2.194529e-20  1.972696e-20  5.095616e-21 -1.123158e-20  3.331811e-21
#> [236]  2.077439e-20 -1.480072e-20  1.127783e-20 -1.181823e-20 -2.646792e-20
#> [241]  5.534034e-21  1.873967e-20  1.008572e-20 -1.114059e-20  4.364664e-21
#> [246] -1.182263e-20 -1.060032e-20  2.527509e-20  1.213903e-20  7.855753e-21
#> [251]  1.730744e-20 -5.362643e-21  1.840626e-20 -1.341522e-20  9.521825e-21
#> [256] -3.687822e-21  8.709974e-21  6.658586e-21 -3.704250e-21 -7.676016e-21
#> [261] -1.590446e-20 -1.636873e-20  4.665688e-22 -1.343333e-20 -1.218679e-20
#> [266]  1.232091e-21 -1.462898e-20 -3.762262e-21  6.328931e-21  1.037726e-20
#> [271] -6.933744e-21 -1.525223e-20  2.871564e-20  1.775882e-20 -2.369761e-21
#> [276]  2.071605e-20  6.923294e-21 -7.343778e-21  1.949168e-20  5.268054e-22
#> [281] -1.249728e-20  2.632982e-21 -5.551638e-21  3.151521e-20 -2.962073e-20
#> [286] -5.339336e-21  1.540012e-20 -9.375912e-21 -6.714567e-21  1.158237e-21
#> [291] -6.401412e-21 -4.190772e-20  2.645166e-20  8.365513e-21 -1.961975e-20
#> [296]  3.977435e-20  4.634902e-22 -9.625087e-21  1.376090e-20  1.898501e-20
#> [301]  3.642783e-21  2.126813e-20  1.611704e-20 -8.912691e-21 -3.567664e-20
#> [306] -2.237904e-20 -1.003170e-20  9.719260e-21  2.222761e-21 -6.416474e-21
#> [311]  3.239194e-21 -4.323131e-21 -6.560791e-21  6.626592e-22 -4.081613e-21
#> [316]  6.646642e-21 -2.892502e-21 -1.719936e-20 -1.466390e-20 -2.076457e-20
#> [321] -1.064754e-20  6.205364e-21  1.002147e-20  4.246898e-22  3.424775e-20
#> [326] -6.625674e-21 -1.444822e-20 -1.011884e-20  2.226544e-20  2.581743e-20
#> [331]  1.679362e-20  8.498285e-21  7.829560e-21 -4.901868e-21  1.392537e-21
#> [336]  3.021435e-20  2.064866e-20 -7.273897e-21  9.988017e-21 -1.752009e-20
#> [341]  6.142438e-21  6.337562e-21 -3.620891e-20 -1.069374e-20  3.921055e-20
#> [346]  1.450656e-20 -2.060520e-20 -1.104800e-20 -1.524043e-20 -8.240455e-21
#> [351] -1.237152e-20 -1.187374e-20 -1.424425e-21  3.180335e-20 -1.929797e-20
#> [356]  6.189839e-21 -2.454710e-20 -1.802275e-20 -2.696358e-21 -2.462481e-21
#> [361] -4.485674e-03  7.536718e-21 -8.655133e-21 -1.729638e-20 -1.857008e-20
#> [366]  2.351535e-20 -1.504682e-20 -5.795512e-21 -4.525498e-21  1.029930e-20
#> [371]  3.856797e-21  2.357215e-20 -1.700043e-20  1.229998e-21  1.335349e-20
#> [376] -3.101955e-20  2.407529e-20 -7.432841e-21  1.654430e-20  5.122707e-21
#> [381] -1.660381e-20 -5.291469e-21  9.322946e-21 -4.299354e-21 -1.284310e-21
#> [386]  3.324351e-21  3.810629e-21  8.385073e-21 -2.688395e-21 -8.265659e-21
#> [391] -3.252199e-20  2.038088e-20  8.231350e-21  2.540457e-20 -1.306083e-20
#> [396] -7.966987e-21 -2.291016e-20  9.484347e-21  1.892535e-20  5.955275e-22
#> [401] -2.044206e-20
#> 
#> $step
#> [1] 2
#>