Inference for linear functional in the high-dimensional linear regression model

Computes the bias corrected estimator of linear functional for the high-dimensional linear regression model and the corresponding standard error.

LF(
  X,
  y,
  loading,
  init.Lasso = NULL,
  lambda = NULL,
  intercept = FALSE,
  mu = NULL,
  step = NULL,
  resol = 1.5,
  maxiter = 10
)

Arguments

X
Design matrix

y
Response variable

loading
observation vector in the linear functional

init.Lasso
initial LASSO estimator of the regression vector (default = NULL)

lambda
Tuning parameter in construction of LASSO estimator of the regression vector (default = NULL)

intercept
Should intercept(s) be fitted (default = FALSE)

mu
Tuning parameter in construction of projection direction (default = NULL)

step
Number of steps (< maxiter) to obtain the smallest mu that gives convergence of the optimization problem for constructing the projection direction (default = NULL)

resol
Resolution or the factor by which mu is increased/decreased to obtain the smallest mu that gives convergence of the optimization problem for constructing the projection direction (default = 1.5)

maxiter
Maximum number of steps along which mu is increased/decreased to obtain the smallest mu that gives convergence of the optimization problem for constructing the projection direction (default = 10)

Value

prop.est

The bias-corrected estimator for the linear functional

sigma

Estimate of the error variance in the linear regression model

se

Standard error of the bias-corrected estimator

proj

Optimal projection direction

step

Number of steps (< maxiter) to obtain the smallest mu that gives convergence of the optimization problem for constructing the projection direction

plug.in

Plug-in LASSO estimator for the linear functional

References

Cai T, Cai TT, Guo Z (2019). “Individualized Treatment Selection: An Optimal Hypothesis Testing Approach In High-dimensional Models.” Unknown. https://statistics.rutgers.edu/home/zijguo/Cai-Cai-Guo-ITE-Selection.pdf.

Examples

X = matrix(sample(-2:2,100*400,replace = TRUE),nrow=100,ncol=400)
beta = (1:400)/25
y = X%*%beta + rnorm(100,0,1)
LF(X = X, y = y, loading = c(1,rep(0,399)), intercept = TRUE)

#> [1] 2
#> [1] "step is 2"

#> $prop.est
#>         [,1]
#> [1,] 44.1424
#> 
#> $sigma
#> [1] 68746.61
#> 
#> $se
#> [1] 6299.181
#> 
#> $proj
#>   [1]  5.151780e-01  3.131494e-01  1.321061e-21  1.606982e-22  2.804835e-22
#>   [6] -4.460346e-22  2.205750e-23 -1.925952e-22 -1.438210e-22 -4.033035e-24
#>  [11]  4.187076e-22  2.463792e-22 -2.567080e-22 -9.744050e-22  2.164287e-22
#>  [16] -3.680118e-22  3.312377e-22  2.793187e-22  6.316993e-23 -5.040983e-22
#>  [21] -7.358200e-22  8.002092e-22  1.880528e-22 -5.548726e-22  1.078487e-21
#>  [26] -2.658169e-22 -2.598480e-22 -8.338432e-22  5.830380e-22  9.017105e-22
#>  [31] -1.942046e-22  4.199569e-22  1.005336e-21 -3.513446e-22 -1.811301e-22
#>  [36] -5.157256e-22  9.723563e-22 -4.272364e-23  1.053911e-21 -3.360194e-22
#>  [41]  1.970230e-22  8.839825e-22 -4.830084e-22 -1.919336e-21  1.527206e-21
#>  [46] -2.860067e-23  3.793679e-22 -2.590791e-22 -3.246531e-22  6.308206e-23
#>  [51] -4.940892e-03 -2.069239e-22  1.747715e-22  1.353585e-21 -5.474266e-22
#>  [56]  1.683669e-22 -1.476235e-02  1.004291e-21  4.545920e-22  1.311543e-21
#>  [61]  7.948699e-22  2.687066e-22 -9.402749e-22 -5.435768e-23 -2.592952e-22
#>  [66]  8.561937e-23  9.282823e-22 -4.194784e-22 -2.941765e-22 -4.026412e-22
#>  [71] -2.155385e-22  4.639235e-22  8.550694e-22 -3.813237e-22  4.841933e-22
#>  [76] -9.809507e-22  9.507746e-23  5.291090e-22 -2.502545e-22 -1.949421e-22
#>  [81] -5.942344e-22 -1.230821e-21 -4.703709e-22  1.366738e-22  4.063641e-22
#>  [86]  4.721058e-22  5.133973e-22 -7.570334e-23  4.093531e-22 -4.417166e-22
#>  [91]  1.275476e-21  2.074675e-23  9.640543e-22  3.365689e-22  1.003069e-22
#>  [96]  7.641882e-22  9.576456e-22 -1.748765e-22  5.665190e-22 -1.102964e-21
#> [101] -5.205661e-22 -1.783353e-22  5.593791e-22  5.348277e-22  4.901336e-22
#> [106] -7.746217e-22 -2.816609e-22  9.102188e-22  5.319321e-22  1.684730e-22
#> [111]  6.990978e-22  3.677703e-22 -6.484807e-22  2.847952e-22 -9.364813e-22
#> [116]  3.153784e-22 -1.829687e-23 -2.837279e-22  7.808408e-22  1.446898e-21
#> [121]  6.508257e-23  3.006437e-22  4.868532e-22  5.013241e-22 -1.755601e-23
#> [126]  8.383819e-22 -8.094963e-23  3.464006e-22  5.854132e-22 -9.322304e-22
#> [131]  6.686827e-22  1.482476e-21 -5.716792e-22  8.977398e-22 -5.131243e-22
#> [136]  5.087354e-22 -5.073014e-22  4.000195e-23 -3.591619e-22 -3.967406e-22
#> [141] -7.664499e-22 -1.228808e-21 -7.213537e-22  6.260619e-22  1.852942e-22
#> [146] -7.871028e-22 -4.374931e-22  1.684407e-21 -1.101795e-21 -4.872193e-22
#> [151] -4.590508e-22 -3.339522e-22  1.093810e-21  1.405747e-21 -2.990940e-22
#> [156] -3.332545e-22  1.401375e-22 -8.386668e-22  2.386198e-23 -3.785107e-23
#> [161]  1.026425e-21  5.658413e-22 -9.484430e-22 -5.643363e-22 -2.392907e-22
#> [166] -7.098773e-22  5.246662e-22 -1.634861e-02 -2.476450e-22  8.346886e-22
#> [171] -1.382135e-21  6.939414e-22 -8.359313e-22 -1.960526e-22 -5.839987e-22
#> [176]  6.187695e-22  3.585767e-22 -4.613540e-22 -1.333510e-21  9.506247e-22
#> [181] -7.970588e-24  1.937715e-22  2.319806e-22 -4.914991e-22 -6.011080e-22
#> [186] -2.703410e-22 -5.033830e-23  7.150530e-22 -2.044594e-22 -6.257336e-22
#> [191]  5.004927e-22  4.896510e-23  1.103673e-22  3.843550e-22 -2.585857e-22
#> [196] -3.883066e-22  7.452543e-22 -8.796755e-22 -9.956033e-23 -8.371020e-23
#> [201] -3.718124e-22  9.733883e-22 -4.684405e-22  3.986411e-22  3.460715e-22
#> [206] -1.678809e-22 -5.895572e-23 -1.697459e-22  2.694112e-22  5.909449e-23
#> [211] -1.162127e-22  8.712325e-22  1.404526e-22 -5.745940e-22 -6.078240e-22
#> [216]  1.488656e-23 -6.894335e-22  1.536548e-23 -2.558156e-22 -4.322175e-22
#> [221] -9.572227e-22 -2.290297e-22  6.495816e-22 -4.035419e-22  3.675762e-03
#> [226] -1.094626e-21 -7.109433e-22 -8.976756e-23  7.563011e-22 -5.153408e-22
#> [231] -1.119003e-21  2.011412e-22  1.227004e-22  5.950427e-23  9.302333e-22
#> [236]  4.611210e-22  1.921559e-22  2.184821e-22  7.276116e-22 -3.992577e-22
#> [241] -1.651434e-21  2.702989e-22  1.752769e-22 -7.796989e-22 -6.921549e-23
#> [246]  4.567892e-22 -2.388774e-22 -4.710694e-22 -7.745541e-22  1.337191e-22
#> [251] -4.667516e-22  1.013343e-21  3.402627e-22  1.585169e-22  1.176199e-21
#> [256] -5.691615e-22  1.258488e-22 -2.114064e-23  2.501259e-23  2.704526e-22
#> [261]  3.402156e-22 -3.844211e-22 -1.060111e-21  5.042880e-22 -4.684587e-23
#> [266]  5.617640e-22  1.438963e-22  3.754825e-22  5.935438e-22  6.765850e-22
#> [271]  5.316814e-22  4.929086e-22 -1.128379e-22  5.998832e-22 -5.429434e-22
#> [276] -2.018014e-02  1.425121e-21  4.794957e-22  7.809421e-22  1.047034e-21
#> [281] -8.826092e-22  7.265825e-22  7.194265e-22  4.083792e-22  1.144482e-21
#> [286] -7.116570e-22 -4.502048e-22  3.923148e-22  1.686902e-23  2.878150e-22
#> [291] -2.479345e-22  6.364569e-22  7.302279e-23  5.171378e-22 -6.510007e-24
#> [296] -3.715849e-22 -4.078399e-22  9.484809e-22 -5.571823e-22  5.983589e-22
#> [301]  3.124719e-24 -5.836613e-22 -3.146346e-22  7.013851e-23  3.284317e-23
#> [306]  9.418624e-23 -2.232158e-22 -6.938510e-23 -4.577275e-22 -2.444216e-22
#> [311]  2.690179e-22 -4.985741e-22 -3.730387e-22 -2.093058e-22 -2.472464e-22
#> [316]  4.490903e-22  1.715352e-22 -7.259815e-22  6.496546e-22  1.387455e-21
#> [321] -4.398210e-22  3.299633e-22  2.874106e-22 -1.629897e-23  1.468401e-22
#> [326] -4.753464e-23  2.298135e-22  7.772295e-22  6.543652e-23 -7.965264e-22
#> [331] -7.871387e-22 -7.139714e-22  4.279113e-22 -7.864276e-22  4.558179e-22
#> [336] -2.224463e-22 -1.902594e-22 -2.068215e-22 -4.710752e-22 -1.453940e-21
#> [341] -4.640233e-22 -2.046968e-22 -1.603809e-22 -7.377653e-22 -2.126465e-22
#> [346] -4.397994e-22  1.449700e-22 -6.638219e-22 -1.369384e-21 -6.787821e-22
#> [351] -1.206444e-21  8.093869e-22 -7.802442e-22 -2.115226e-22 -4.116403e-22
#> [356] -5.636029e-22  5.750233e-22 -7.954935e-22  1.488206e-22  2.590991e-22
#> [361] -2.759258e-23 -1.231083e-22  1.186965e-21 -9.719618e-22 -8.476994e-22
#> [366]  8.576943e-22 -5.093589e-22  7.259530e-23  4.927049e-24 -7.430229e-22
#> [371] -1.256530e-22  6.018740e-22 -1.053666e-21  5.566682e-22  4.203998e-22
#> [376]  2.036601e-22  7.859942e-22  3.222071e-22 -3.318526e-22  3.213144e-22
#> [381]  1.804583e-22  6.459827e-22 -1.224084e-02 -4.579500e-22 -6.399157e-22
#> [386] -1.019241e-21  4.687759e-22  3.899103e-22 -5.764814e-22 -4.648942e-22
#> [391]  8.362310e-22 -9.560480e-22  4.665031e-22  7.092873e-22 -6.406239e-22
#> [396]  1.179475e-21 -2.508748e-22  2.911556e-22 -7.650071e-22 -6.378236e-22
#> [401]  6.834668e-23
#> 
#> $step
#> [1] 2
#> 
#> $plug.in
#> [1] 27.34158
#>