Inference for linear functional in the high-dimensional linear regression model

Computes the bias corrected estimator of linear functional for the high-dimensional linear regression model and the corresponding standard error.

LF_Inference(
  X,
  y,
  loading,
  init.Lasso = NULL,
  lambda = NULL,
  intercept = FALSE,
  mu = NULL,
  step = NULL,
  resol = 1.5,
  maxiter = 10
)

Arguments

X
Design matrix

y
Response variable

loading
observation vector in the linear functional

init.Lasso
initial LASSO estimator of the regression vector (default = NULL)

lambda
Tuning parameter in construction of LASSO estimator of the regression vector (default = NULL)

intercept
Should intercept(s) be fitted (default = FALSE)

mu
Tuning parameter in construction of projection direction (default = NULL)

step
Number of steps (< maxiter) to obtain the smallest mu that gives convergence of the optimization problem for constructing the projection direction (default = NULL)

resol
Resolution or the factor by which mu is increased/decreased to obtain the smallest mu that gives convergence of the optimization problem for constructing the projection direction (default = 1.5)

maxiter
Maximum number of steps along which mu is increased/decreased to obtain the smallest mu that gives convergence of the optimization problem for constructing the projection direction (default = 10)

Value

prop.est

The bias-corrected estimator for the linear functional

sigma

Estimate of the error variance in the linear regression model

se

Standard error of the bias-corrected estimator

proj

Optimal projection direction

step

Number of steps (< maxiter) to obtain the smallest mu that gives convergence of the optimization problem for constructing the projection direction

plug.in

Plug-in LASSO estimator for the linear functional

References

Cai T, Cai TT, Guo Z (2019). “Individualized Treatment Selection: An Optimal Hypothesis Testing Approach In High-dimensional Models.” Unknown. https://statistics.rutgers.edu/home/zijguo/Cai-Cai-Guo-ITE-Selection.pdf.

Examples

X = matrix(sample(-2:2,100*400,replace = TRUE),nrow=100,ncol=400)
beta = (1:400)/25
y = X%*%beta + rnorm(100,0,1)
LF_Inference(X = X, y = y, loading = c(1,rep(0,399)), intercept = TRUE)

#> [1] 1
#> [1] 0.1257833
#> [1] 2
#> [1] 0.1663815
#> [1] 3
#> [1] 0.2299059
#> [1] 4
#> [1] 3
#> [1] "step is 3"

#> $prop.est
#>          [,1]
#> [1,] 35.72155
#> 
#> $sigma
#> [1] 70149.6
#> 
#> $se
#> [1] 8063.904
#> 
#> $proj
#>                 [,1]
#>   [1,]  2.247013e+00
#>   [2,]  1.122457e+00
#>   [3,] -2.266709e-19
#>   [4,] -3.898583e-19
#>   [5,]  8.648411e-19
#>   [6,] -3.632335e-19
#>   [7,] -7.409506e-19
#>   [8,] -6.168825e-19
#>   [9,]  1.152037e-18
#>  [10,]  8.844823e-19
#>  [11,]  1.685311e-01
#>  [12,]  2.017774e-02
#>  [13,]  8.814912e-19
#>  [14,]  6.545094e-19
#>  [15,]  5.266949e-19
#>  [16,] -1.412859e-19
#>  [17,] -8.054976e-19
#>  [18,] -6.104751e-19
#>  [19,] -1.687272e-18
#>  [20,] -1.413293e-18
#>  [21,] -1.003573e-01
#>  [22,]  7.211665e-02
#>  [23,] -1.236845e-01
#>  [24,] -5.189344e-19
#>  [25,]  1.318407e-01
#>  [26,] -5.637378e-19
#>  [27,] -6.981360e-19
#>  [28,]  3.282471e-19
#>  [29,] -3.803952e-19
#>  [30,] -1.781099e-20
#>  [31,]  4.755318e-02
#>  [32,]  9.616470e-20
#>  [33,] -5.298430e-19
#>  [34,] -6.551233e-20
#>  [35,]  5.511071e-19
#>  [36,] -1.853385e-19
#>  [37,]  6.030719e-03
#>  [38,] -1.103679e-18
#>  [39,] -1.080025e-19
#>  [40,]  1.793242e-19
#>  [41,]  3.674408e-02
#>  [42,]  6.016367e-19
#>  [43,]  3.497271e-19
#>  [44,] -1.333342e-01
#>  [45,]  8.866429e-02
#>  [46,]  4.369577e-19
#>  [47,]  6.765803e-19
#>  [48,] -6.639826e-19
#>  [49,]  5.558910e-19
#>  [50,]  9.253176e-02
#>  [51,] -1.196562e-01
#>  [52,] -3.255783e-19
#>  [53,]  1.661488e-19
#>  [54,]  2.158396e-19
#>  [55,] -8.431520e-20
#>  [56,] -4.173436e-19
#>  [57,] -2.346135e-01
#>  [58,] -6.510923e-19
#>  [59,] -5.043573e-19
#>  [60,] -9.913855e-19
#>  [61,] -7.805603e-19
#>  [62,]  3.615050e-19
#>  [63,]  4.107948e-19
#>  [64,]  1.217056e-18
#>  [65,]  4.189211e-19
#>  [66,]  2.857953e-02
#>  [67,] -8.529205e-19
#>  [68,]  9.139680e-19
#>  [69,]  2.222420e-19
#>  [70,] -1.959436e-19
#>  [71,]  8.379308e-19
#>  [72,] -2.616791e-19
#>  [73,]  3.504476e-20
#>  [74,]  9.265599e-19
#>  [75,] -1.858413e-18
#>  [76,]  6.467783e-19
#>  [77,]  4.332802e-19
#>  [78,]  1.752465e-01
#>  [79,] -1.015782e-19
#>  [80,] -4.129590e-19
#>  [81,]  7.810712e-19
#>  [82,] -2.733576e-19
#>  [83,]  2.122530e-19
#>  [84,] -9.244721e-19
#>  [85,] -5.904225e-19
#>  [86,] -6.673735e-19
#>  [87,]  1.130098e-01
#>  [88,]  2.735826e-19
#>  [89,] -7.832785e-20
#>  [90,]  8.790584e-19
#>  [91,]  1.250613e-01
#>  [92,]  8.045081e-19
#>  [93,]  3.115828e-19
#>  [94,] -5.728320e-20
#>  [95,]  2.890226e-19
#>  [96,] -9.063840e-19
#>  [97,] -1.291347e-18
#>  [98,] -6.462734e-19
#>  [99,] -2.869869e-19
#> [100,] -1.004417e-01
#> [101,]  1.483301e-19
#> [102,] -1.481931e-18
#> [103,] -4.350229e-19
#> [104,] -1.490779e-19
#> [105,] -7.635905e-19
#> [106,]  4.402300e-19
#> [107,] -8.029551e-19
#> [108,]  2.899322e-01
#> [109,] -8.221510e-19
#> [110,]  1.157519e-19
#> [111,]  2.385597e-19
#> [112,] -6.082670e-19
#> [113,] -3.992531e-19
#> [114,]  2.813424e-20
#> [115,]  1.577544e-19
#> [116,]  6.128949e-19
#> [117,] -1.859360e-19
#> [118,]  1.087670e-19
#> [119,]  4.405990e-19
#> [120,]  3.822129e-19
#> [121,] -3.155036e-19
#> [122,] -7.732558e-19
#> [123,]  1.089903e-19
#> [124,] -2.404219e-19
#> [125,]  6.717505e-20
#> [126,]  6.334457e-19
#> [127,]  3.049299e-19
#> [128,]  9.470518e-02
#> [129,]  5.217266e-19
#> [130,] -1.188223e-18
#> [131,]  9.862536e-19
#> [132,]  2.663075e-21
#> [133,] -1.256391e-18
#> [134,]  8.189886e-19
#> [135,]  9.814163e-20
#> [136,]  8.145775e-19
#> [137,]  1.678579e-19
#> [138,] -1.613430e-19
#> [139,] -3.461842e-02
#> [140,] -4.191449e-02
#> [141,]  8.613356e-19
#> [142,]  9.525776e-20
#> [143,]  3.895523e-19
#> [144,]  2.925591e-19
#> [145,] -6.589670e-19
#> [146,]  3.485461e-20
#> [147,] -1.104820e-18
#> [148,]  7.681879e-02
#> [149,] -4.603533e-02
#> [150,] -2.015209e-19
#> [151,]  5.531877e-19
#> [152,] -3.662427e-19
#> [153,]  8.748156e-19
#> [154,]  1.261439e-19
#> [155,]  9.303262e-20
#> [156,] -1.221038e-18
#> [157,] -6.033877e-19
#> [158,] -8.307521e-19
#> [159,] -3.270863e-19
#> [160,] -3.238734e-19
#> [161,]  1.038675e-01
#> [162,]  1.296869e-19
#> [163,]  1.719084e-18
#> [164,] -3.213235e-19
#> [165,] -1.592306e-02
#> [166,] -7.526468e-19
#> [167,]  2.689380e-02
#> [168,] -1.773854e-01
#> [169,]  8.080709e-19
#> [170,]  8.020866e-19
#> [171,] -1.059502e-01
#> [172,]  2.251430e-19
#> [173,]  5.561855e-19
#> [174,] -7.582608e-20
#> [175,]  2.099994e-19
#> [176,] -1.570261e-18
#> [177,] -1.015579e-18
#> [178,]  7.662487e-20
#> [179,]  1.660584e-19
#> [180,] -5.005474e-20
#> [181,]  1.876709e-19
#> [182,]  1.652840e-19
#> [183,] -1.984627e-19
#> [184,]  4.189377e-19
#> [185,] -2.355941e-19
#> [186,]  1.028908e-18
#> [187,] -8.800740e-19
#> [188,] -5.283637e-19
#> [189,]  7.644176e-20
#> [190,]  7.519078e-19
#> [191,] -4.131462e-20
#> [192,] -1.219730e-18
#> [193,]  2.800851e-19
#> [194,]  1.056665e-18
#> [195,]  1.590832e-19
#> [196,]  4.806366e-19
#> [197,] -1.434915e-18
#> [198,]  1.233789e-18
#> [199,]  3.523082e-02
#> [200,] -3.805192e-19
#> [201,] -2.578359e-19
#> [202,]  7.181162e-19
#> [203,]  1.131054e-19
#> [204,] -4.446256e-19
#> [205,]  5.740337e-19
#> [206,]  1.687553e-20
#> [207,] -4.763246e-20
#> [208,] -5.397460e-02
#> [209,]  6.189937e-02
#> [210,]  1.625723e-19
#> [211,]  9.884452e-19
#> [212,]  5.474837e-19
#> [213,] -1.391288e-18
#> [214,]  3.357533e-19
#> [215,] -2.049894e-19
#> [216,] -1.458719e-19
#> [217,] -5.015830e-03
#> [218,] -5.286398e-19
#> [219,] -1.311355e-19
#> [220,] -1.169942e-01
#> [221,] -5.583221e-20
#> [222,]  2.156313e-19
#> [223,] -1.358975e-19
#> [224,] -5.657889e-19
#> [225,]  3.932259e-01
#> [226,] -2.463292e-01
#> [227,] -2.476852e-01
#> [228,] -3.574490e-19
#> [229,] -6.469600e-19
#> [230,]  4.221361e-19
#> [231,]  1.507528e-19
#> [232,]  1.533714e-18
#> [233,]  3.625147e-19
#> [234,]  1.773354e-19
#> [235,]  2.590352e-01
#> [236,] -4.625634e-19
#> [237,]  1.353844e-20
#> [238,] -1.808388e-19
#> [239,] -5.699141e-19
#> [240,] -2.496482e-20
#> [241,] -1.403909e-01
#> [242,] -9.056785e-19
#> [243,]  1.632305e-02
#> [244,] -2.458800e-20
#> [245,]  1.045289e-18
#> [246,] -8.485000e-19
#> [247,]  5.252549e-23
#> [248,] -6.836028e-19
#> [249,] -2.010142e-19
#> [250,]  3.497914e-02
#> [251,] -1.183850e-18
#> [252,]  1.259784e-18
#> [253,] -5.087350e-19
#> [254,]  1.884889e-02
#> [255,]  1.568709e-18
#> [256,]  2.072892e-19
#> [257,] -8.175840e-19
#> [258,] -1.135559e-18
#> [259,] -6.251465e-19
#> [260,]  8.190219e-21
#> [261,]  2.247211e-19
#> [262,] -8.895269e-02
#> [263,] -8.427835e-02
#> [264,] -9.254963e-19
#> [265,] -2.341433e-19
#> [266,] -6.707999e-20
#> [267,] -1.313843e-18
#> [268,]  3.374828e-19
#> [269,] -6.559537e-19
#> [270,]  8.533641e-19
#> [271,]  3.972534e-19
#> [272,] -3.902867e-19
#> [273,] -1.095973e-18
#> [274,]  5.175340e-19
#> [275,]  1.025044e-18
#> [276,] -3.485689e-01
#> [277,]  8.462411e-02
#> [278,] -1.693107e-20
#> [279,]  3.587038e-19
#> [280,] -7.053696e-20
#> [281,] -1.034949e-01
#> [282,] -5.252622e-20
#> [283,] -3.981524e-19
#> [284,]  1.073222e-19
#> [285,] -1.630978e-18
#> [286,]  2.077192e-19
#> [287,] -1.444337e-01
#> [288,] -2.191601e-19
#> [289,]  4.011667e-19
#> [290,]  4.249313e-19
#> [291,] -6.348256e-20
#> [292,] -7.247518e-19
#> [293,] -1.038600e-18
#> [294,] -8.158762e-19
#> [295,]  2.359100e-19
#> [296,]  3.618594e-19
#> [297,] -7.117540e-19
#> [298,]  2.263342e-19
#> [299,]  5.370467e-19
#> [300,] -7.443262e-19
#> [301,]  4.393052e-19
#> [302,] -4.922723e-02
#> [303,] -9.515861e-19
#> [304,]  1.113995e-18
#> [305,] -1.418110e-18
#> [306,] -1.285751e-18
#> [307,]  3.259700e-19
#> [308,] -6.973371e-19
#> [309,]  2.020424e-19
#> [310,] -7.760566e-19
#> [311,] -1.117488e-18
#> [312,]  7.123819e-20
#> [313,]  1.072944e-18
#> [314,] -1.036153e-19
#> [315,]  3.745801e-19
#> [316,]  1.613875e-19
#> [317,] -7.531601e-19
#> [318,]  8.107332e-19
#> [319,] -7.565366e-20
#> [320,]  3.445795e-01
#> [321,]  4.683752e-02
#> [322,]  5.842222e-20
#> [323,]  1.892179e-18
#> [324,] -1.912900e-18
#> [325,] -4.921650e-19
#> [326,] -1.149804e-18
#> [327,] -6.560281e-19
#> [328,] -4.074312e-19
#> [329,] -5.288394e-19
#> [330,] -7.187312e-19
#> [331,] -2.193930e-01
#> [332,] -2.786521e-02
#> [333,] -6.213877e-20
#> [334,]  2.861697e-19
#> [335,]  3.038075e-02
#> [336,] -4.903889e-19
#> [337,]  7.752899e-19
#> [338,] -8.862484e-19
#> [339,]  2.503782e-19
#> [340,]  8.078122e-21
#> [341,] -5.805432e-02
#> [342,]  2.173432e-19
#> [343,] -1.003063e-18
#> [344,] -1.338661e-01
#> [345,]  2.267428e-19
#> [346,]  3.499847e-19
#> [347,]  8.185289e-20
#> [348,] -1.564631e-19
#> [349,]  5.384980e-19
#> [350,]  1.202739e-19
#> [351,] -2.019690e-02
#> [352,] -4.631605e-19
#> [353,] -2.971770e-19
#> [354,] -6.081011e-19
#> [355,]  2.451476e-19
#> [356,]  3.948406e-19
#> [357,]  7.378295e-20
#> [358,] -9.469765e-20
#> [359,] -1.219263e-18
#> [360,]  2.240501e-19
#> [361,]  3.433476e-19
#> [362,] -8.277623e-19
#> [363,] -9.180769e-19
#> [364,] -2.060674e-01
#> [365,]  1.113900e-19
#> [366,] -4.683858e-19
#> [367,] -8.421520e-19
#> [368,]  4.717519e-19
#> [369,]  1.516669e-18
#> [370,] -6.296665e-19
#> [371,] -5.001190e-19
#> [372,]  3.090653e-19
#> [373,]  6.335920e-19
#> [374,]  4.086076e-20
#> [375,] -1.482519e-19
#> [376,]  1.110487e-01
#> [377,]  1.086788e-01
#> [378,]  2.631067e-02
#> [379,] -1.886331e-19
#> [380,] -9.962035e-19
#> [381,]  1.019129e-18
#> [382,]  4.421136e-20
#> [383,] -4.141540e-01
#> [384,] -5.961603e-19
#> [385,] -9.043522e-02
#> [386,] -1.040162e-01
#> [387,] -8.627369e-19
#> [388,]  3.005570e-19
#> [389,]  2.066603e-19
#> [390,]  3.919803e-19
#> [391,]  1.232283e-18
#> [392,] -2.305763e-01
#> [393,] -1.234700e-18
#> [394,]  9.198298e-19
#> [395,] -1.131918e-18
#> [396,]  2.340758e-02
#> [397,]  4.233068e-19
#> [398,]  5.405736e-19
#> [399,]  5.306909e-19
#> [400,] -3.128533e-02
#> [401,] -9.925488e-02
#> 
#> $step
#> [1] 3
#> 
#> $plug.in
#> [1] 27.34158
#>