Inference for linear functional in the high-dimensional logistic regression model

Computes the bias corrected estimator of the linear functional for the high-dimensional logistic regression model and the corresponding standard error.

LF_Inference_logistic(
  X,
  y,
  loading,
  weight,
  lambda = NULL,
  intercept = FALSE,
  mu = NULL,
  step = NULL,
  resol = 1.5,
  maxiter = 6
)

Arguments

X
Design matrix

y
Response vector

loading
observation vector in the linear functional

weight
The vector of weights used in correcting the plug-in estimator

lambda
Tuning parameter in construction of the LASSO estimator of the regression vector (default = NULL)

intercept
Should intercept(s) be fitted (default = FALSE)

mu
Tuning parameter in construction of the projection direction (default = NULL)

step
Number of steps (< maxiter) to obtain the smallest mu that gives convergence of the optimization problem for constructing the projection direction (default = NULL)

resol
Resolution or the factor by which mu is increased/decreased to obtain the smallest mu that gives convergence of the optimization problem for constructing the projection direction (default = 1.5)

maxiter
Maximum number of steps along which mu is increased/decreased to obtain the smallest mu that gives convergence of the optimization problem for constructing the projection direction (default = 6)

Value

prop.est

The bias corrected estimator of the linear functional

se

Standard error of the bias-corrected estimator

proj

Optimal projection direction

step

Number of steps (< maxiter) to obtain the smallest mu that gives convergence of the optimization problem for constructing the projection direction

plug.in

Plug-in LASSO estimator of the linear functional

support

Estimated support for the regression vector (using LASSO estimate)

References

Guo Z, Rakshit P, Herman DS, Chen J (2019). “Inference for Case Probability in High-dimensional Logistic Regression.” Unknown.

Examples

LF_Inference_logistic(X = matrix(sample(-2:2,50*300,replace = TRUE),nrow=50,ncol=300),
                      y = rbinom(50,1,0.5), loading = c(1,rep(0,299)),
                      intercept = TRUE, weight = rep(1,50))

#> [1] 2
#> [1] "step is 2"

#> $prop.est
#> [1] 0.7006761
#> 
#> $se
#> [1] 0.2662161
#> 
#> $proj
#>   [1]  2.173891e+00  1.655794e+00 -1.058876e-19 -4.258436e-20  6.666689e-20
#>   [6]  1.070594e-19 -4.590542e-20 -4.974499e-20  9.900552e-20 -5.207874e-21
#>  [11] -5.197358e-20  4.952432e-20  1.283698e-19 -3.644112e-20 -1.394551e-19
#>  [16]  1.577036e-20  9.375502e-20 -3.300569e-21 -2.167718e-21 -1.917185e-20
#>  [21]  1.250134e-20  6.608010e-20  3.330395e-20 -3.836143e-20  6.596529e-20
#>  [26]  2.605006e-20 -4.149445e-20  3.513932e-20 -5.921570e-20 -1.204898e-19
#>  [31]  1.506660e-19  5.847998e-20 -1.237953e-19 -1.278813e-19 -2.196291e-20
#>  [36]  6.733131e-20 -1.134309e-19 -6.694687e-20 -5.501430e-21  2.775297e-19
#>  [41]  9.552091e-20 -1.575403e-19  9.662984e-20  3.402942e-20  1.574531e-19
#>  [46]  7.621477e-20 -6.001735e-20  7.454932e-21 -5.425387e-20 -7.858246e-20
#>  [51] -2.098261e-20  3.181506e-19 -5.245711e-20  5.703869e-20 -2.119470e-19
#>  [56] -6.083341e-20  1.671849e-19 -1.034859e-19 -4.539118e-20 -5.786762e-20
#>  [61]  4.490209e-20  1.442476e-19  9.567267e-20 -8.450305e-20 -2.316189e-20
#>  [66] -1.053820e-19  3.092365e-20  8.201887e-20  6.324813e-20  9.311696e-20
#>  [71] -5.426317e-20 -4.226375e-20 -5.171762e-20  1.798664e-19 -4.819633e-20
#>  [76] -2.362343e-20  1.173445e-19  8.171087e-20  3.935218e-20  6.118351e-20
#>  [81]  2.530178e-20 -1.053810e-19 -6.759080e-20 -1.873174e-19  3.245495e-20
#>  [86] -4.440171e-20 -1.084689e-19 -4.774199e-20  5.976092e-20  2.232102e-20
#>  [91]  1.816820e-20  4.864407e-20 -3.012039e-20  1.399257e-20  2.733829e-20
#>  [96]  1.242985e-19  7.195635e-20  1.945624e-20  1.198724e-19 -6.386127e-20
#> [101]  1.171759e-20 -1.302210e-19 -6.943513e-20  1.806734e-19 -1.567344e-20
#> [106] -1.513708e-19  1.286529e-20  1.813798e-19  9.290906e-21  2.040480e-19
#> [111]  4.331930e-20 -1.636534e-19  1.557268e-19  1.172450e-19 -3.246265e-20
#> [116]  1.374269e-19 -2.403029e-19  1.089003e-19  3.278710e-20  5.397209e-20
#> [121] -4.981777e-20  6.117102e-20  1.869754e-19  2.181011e-19 -2.704737e-19
#> [126] -1.841433e-20 -1.017598e-19 -1.376572e-19 -9.647422e-20  2.501836e-21
#> [131]  9.335158e-20  1.186879e-19 -5.853334e-20  1.762878e-19 -1.482073e-19
#> [136] -1.120284e-19 -1.718277e-20  6.473845e-20  1.728398e-19 -1.959235e-21
#> [141]  1.195087e-19 -9.674352e-20  3.056018e-20 -1.314885e-19 -2.324317e-20
#> [146]  5.489923e-21 -8.435200e-20 -1.730460e-19  2.321073e-19 -4.393951e-20
#> [151]  9.027939e-20  1.215460e-19  7.754706e-20  1.223963e-19  7.139832e-20
#> [156]  5.834645e-20 -1.367435e-20  9.544853e-20  2.677508e-19 -1.663539e-19
#> [161] -8.637528e-20 -1.557084e-19  2.119871e-19 -3.467510e-21  2.435195e-19
#> [166]  1.540653e-19 -1.890262e-19  9.285277e-20 -7.282589e-20  2.321578e-21
#> [171] -1.568364e-19  1.653375e-20 -1.883132e-20  4.603858e-21  1.277655e-19
#> [176] -9.820117e-20  4.103275e-20  1.274485e-19 -8.060724e-21  1.808885e-20
#> [181]  1.499251e-19 -5.155910e-20 -4.057822e-20 -1.712007e-19 -5.806557e-20
#> [186]  1.751655e-19 -5.530585e-20  3.093623e-21 -9.202534e-20  2.779978e-20
#> [191]  1.752040e-19 -9.136359e-20 -2.444435e-20  1.393218e-20  1.159094e-20
#> [196] -3.494062e-20 -2.020163e-19  1.962645e-20 -1.431956e-19 -1.326555e-19
#> [201]  3.497435e-20  6.515794e-20 -1.217150e-19 -6.729593e-22  2.545868e-19
#> [206]  1.434597e-19 -1.477234e-19  4.490328e-20 -4.082942e-20  1.158412e-19
#> [211]  4.502129e-02  2.196376e-19 -1.423467e-19 -3.305466e-20  1.012788e-20
#> [216] -1.134182e-19  2.470720e-20  8.126982e-20 -1.326865e-19 -2.530022e-21
#> [221] -8.428113e-20  5.083006e-20  8.684688e-20  4.132361e-20 -2.065082e-20
#> [226]  3.284630e-20 -2.527981e-19 -7.305113e-20  1.780654e-19  1.024300e-19
#> [231] -1.036904e-19  4.316458e-20 -1.395614e-19  1.738086e-19  1.713938e-19
#> [236]  9.004933e-20  1.918706e-19  1.248014e-19 -8.509998e-20  8.653851e-20
#> [241] -1.495216e-19 -9.045218e-20 -1.076385e-19  1.245888e-19  1.197616e-19
#> [246] -1.552396e-19  2.677958e-20 -6.845505e-20  1.156951e-20  1.240116e-19
#> [251] -1.424647e-19 -1.394862e-19 -1.187461e-20  7.300007e-21  5.702674e-20
#> [256] -2.608138e-20 -8.484511e-20  6.372870e-20 -9.676248e-21  1.530345e-19
#> [261]  7.887390e-20 -3.095019e-20 -1.320104e-19 -1.479289e-19 -8.795959e-20
#> [266] -1.021391e-20 -4.053101e-21  1.855927e-21 -1.513726e-19 -2.545829e-20
#> [271]  6.777169e-21  1.844972e-19 -1.649968e-20 -1.123806e-19  7.559906e-20
#> [276]  8.597301e-20 -1.680143e-20  2.039471e-20  1.840689e-19 -1.237614e-20
#> [281]  2.710189e-20  1.779040e-20  2.194943e-20 -1.535927e-19  2.022092e-20
#> [286] -7.142056e-20 -2.167789e-20  7.110259e-20 -4.352966e-20 -1.112371e-19
#> [291]  2.432645e-21 -2.696481e-19  8.125888e-20  2.712968e-02 -7.327839e-20
#> [296]  1.162597e-19  9.457073e-20  2.320533e-19  1.663909e-19  1.108286e-19
#> [301]  8.962311e-20
#> 
#> $step
#> [1] 2
#> 
#> $plug.in
#> [1] 0.4559511
#> 
#> $support
#>   [1]  TRUE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [13] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE
#>  [25] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [37] FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE
#>  [49] FALSE  TRUE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE  TRUE FALSE FALSE
#>  [61] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE
#>  [73] FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [85] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [97] FALSE FALSE  TRUE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE
#> [109] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [121] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [133] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE
#> [145] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [157] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [169] FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE  TRUE
#> [181] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [193] FALSE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [205] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [217] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [229] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [241] FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [253] FALSE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [265] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE
#> [277] FALSE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [289] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [301] FALSE
#>