Inference for linear functional in the high-dimensional logistic regression model

Computes the bias corrected estimator of the linear functional for the high-dimensional logistic regression model and the corresponding standard error.

LF_logistic(
  X,
  y,
  loading,
  weight,
  lambda = NULL,
  intercept = FALSE,
  mu = NULL,
  step = NULL,
  resol = 1.5,
  maxiter = 6
)

Arguments

X
Design matrix

y
Response vector

loading
observation vector in the linear functional

weight
The vector of weights used in correcting the plug-in estimator

lambda
Tuning parameter in construction of the LASSO estimator of the regression vector (default = NULL)

intercept
Should intercept(s) be fitted (default = FALSE)

mu
Tuning parameter in construction of the projection direction (default = NULL)

step
Number of steps (< maxiter) to obtain the smallest mu that gives convergence of the optimization problem for constructing the projection direction (default = NULL)

resol
Resolution or the factor by which mu is increased/decreased to obtain the smallest mu that gives convergence of the optimization problem for constructing the projection direction (default = 1.5)

maxiter
Maximum number of steps along which mu is increased/decreased to obtain the smallest mu that gives convergence of the optimization problem for constructing the projection direction (default = 6)

Value

prop.est

The bias corrected estimator of the linear functional

se

Standard error of the bias-corrected estimator

proj

Optimal projection direction

step

Number of steps (< maxiter) to obtain the smallest mu that gives convergence of the optimization problem for constructing the projection direction

plug.in

Plug-in LASSO estimator of the linear functional

support

Estimated support for the regression vector (using LASSO estimate)

References

Guo Z, Rakshit P, Herman DS, Chen J (2019). “Inference for Case Probability in High-dimensional Logistic Regression.” Unknown.

Examples

LF_logistic(X = matrix(sample(-2:2,50*300,replace = TRUE),nrow=50,ncol=300),
                      y = rbinom(50,1,0.5), loading = c(1,rep(0,299)),
                      intercept = TRUE, weight = rep(1,50))

#> [1] 2
#> [1] "step is 2"

#> $prop.est
#> [1] 0.2812116
#> 
#> $se
#> [1] 0.2403733
#> 
#> $proj
#>   [1]  1.730114e+00  1.464229e+00  1.096121e-20 -2.380769e-20 -1.527051e-20
#>   [6]  1.683891e-20  2.452025e-20 -2.931417e-20 -1.305886e-20 -2.084379e-20
#>  [11] -2.074505e-20  4.175312e-20  2.096838e-20 -3.797969e-20  3.021915e-21
#>  [16]  4.064425e-21 -1.579717e-20 -1.450474e-21 -1.809581e-20 -4.052890e-20
#>  [21] -3.978002e-20 -2.400746e-20 -2.738445e-20  4.578886e-20  9.977090e-21
#>  [26]  2.053183e-20  1.075832e-20 -1.664695e-20 -6.907292e-21 -1.336949e-20
#>  [31]  3.705863e-20 -1.841321e-20 -2.518722e-20  8.154728e-22 -7.227225e-20
#>  [36]  2.068253e-21 -3.628896e-20  4.247174e-21 -2.009011e-20  2.889834e-21
#>  [41]  7.209901e-21  9.362476e-22  3.016039e-21 -5.794124e-20 -3.980828e-20
#>  [46]  2.906863e-20 -1.738548e-20 -1.043501e-21  2.688867e-21 -1.102794e-20
#>  [51] -1.017493e-20 -2.141121e-20  2.927253e-21 -2.549707e-21  1.124738e-21
#>  [56] -1.022523e-20 -6.888658e-20 -4.824587e-20  4.473545e-20  3.355398e-20
#>  [61] -6.257550e-21 -1.537511e-20  1.185588e-20  2.112652e-20  2.217205e-20
#>  [66] -4.117983e-21  5.743067e-21 -1.769691e-20  1.535179e-21 -4.066319e-21
#>  [71]  3.434802e-20 -4.057328e-24 -3.442056e-20  2.409564e-20 -4.524828e-21
#>  [76]  4.598946e-21  3.306309e-20  3.152672e-21  4.481027e-20  7.193150e-21
#>  [81]  5.288364e-20  1.327498e-20  1.971933e-20 -4.718107e-21  1.090656e-20
#>  [86] -2.735860e-20  8.561333e-21  4.198696e-21  3.717980e-20  3.799960e-20
#>  [91] -9.158241e-21  5.114089e-20 -1.376716e-20 -2.096612e-20  1.565873e-20
#>  [96]  3.717051e-20  2.836295e-20 -1.712364e-20 -2.845006e-21  3.230361e-03
#> [101]  1.425009e-20  2.968231e-20  5.446642e-20  1.648545e-20 -1.878245e-20
#> [106]  7.720120e-20 -1.255839e-20  3.113200e-20  2.500681e-20  3.204555e-20
#> [111]  2.705325e-20 -1.701039e-20  9.783250e-21  4.522307e-21 -1.873712e-20
#> [116] -6.632789e-21  4.172915e-20  1.656890e-20  4.029207e-20  8.612210e-21
#> [121] -1.148929e-20 -3.218393e-20 -2.293802e-20 -4.836537e-20 -1.844283e-20
#> [126] -1.366058e-20  2.526139e-20 -1.547570e-20  1.457807e-20  5.830897e-20
#> [131] -2.736274e-20 -3.346888e-20 -4.898152e-20 -3.550726e-20  3.068154e-20
#> [136]  2.857463e-20  3.210360e-20 -1.208583e-20 -9.117168e-21  2.230831e-20
#> [141] -4.230916e-20 -9.654725e-21 -7.344001e-21  2.561686e-20 -1.058455e-20
#> [146] -3.611837e-21 -1.477016e-20 -4.785066e-22 -6.174826e-21 -1.547889e-20
#> [151]  1.227458e-20 -3.093684e-20  5.557875e-21  1.452940e-20  1.858453e-20
#> [156]  8.829611e-21 -1.441908e-20 -1.864904e-20 -1.608913e-20 -3.822988e-20
#> [161]  1.343835e-20 -2.548175e-21  3.973804e-20 -6.399285e-21 -7.238920e-21
#> [166] -1.543264e-20 -1.800479e-20  5.978217e-21  1.323098e-21  2.670660e-22
#> [171] -1.269510e-20  7.441164e-20  2.797970e-20  2.487229e-20  5.602159e-21
#> [176] -3.815496e-21  1.302672e-20  1.234467e-20 -9.183404e-21  4.420886e-20
#> [181]  2.635929e-21  3.052517e-20 -8.359521e-21  6.826724e-21  1.346749e-20
#> [186]  1.402858e-20 -5.144377e-20 -1.755700e-20 -2.157976e-23 -3.593770e-21
#> [191]  1.368244e-20  1.034554e-20 -3.800812e-21  4.253662e-21  2.476814e-20
#> [196] -2.194797e-20  2.138628e-20 -2.207539e-20 -2.980434e-20 -1.044476e-20
#> [201] -1.530841e-21 -1.326061e-21 -1.124654e-20  2.798017e-21  1.753459e-21
#> [206] -8.974915e-21 -1.048871e-20 -4.862268e-20 -4.067999e-21  2.620372e-22
#> [211]  2.586704e-20 -5.069810e-20 -2.655717e-20 -5.236707e-20 -2.262514e-20
#> [216]  8.192117e-21  2.828285e-21 -1.147175e-20 -3.387834e-20  1.183914e-20
#> [221] -1.117093e-20  2.034068e-20 -9.948117e-22  1.796190e-20  3.054913e-20
#> [226] -2.374410e-20 -1.289783e-20 -6.209266e-21 -2.178571e-20 -1.260326e-20
#> [231]  1.537623e-20 -2.092358e-21 -3.547747e-20  7.877986e-21  6.544562e-21
#> [236] -3.774372e-20  1.727900e-20 -5.168015e-21 -4.452928e-21  1.113055e-20
#> [241] -1.561679e-20  2.101050e-01  5.749813e-21  5.866120e-21 -8.764909e-21
#> [246]  1.692944e-20  4.032988e-20 -3.608811e-20 -3.115150e-21 -3.047611e-20
#> [251] -1.069405e-20  4.533990e-21 -3.708255e-20  2.868477e-20  1.596561e-20
#> [256]  6.071692e-22  5.537169e-20 -2.251409e-20  1.028242e-20  3.887744e-20
#> [261]  3.184066e-20  5.996695e-21 -1.176205e-23 -1.066039e-20 -1.669490e-20
#> [266] -2.462409e-20  1.344886e-20 -2.576317e-21 -4.004926e-20 -1.048408e-20
#> [271]  4.231298e-20  4.673691e-21  1.287166e-20  1.841775e-20  1.791926e-20
#> [276] -5.005941e-21 -3.654816e-20  1.960425e-20 -2.147327e-20  3.593999e-20
#> [281] -3.456575e-20  2.263470e-20  1.206497e-20 -4.656906e-20 -3.927503e-20
#> [286] -6.794266e-22 -1.223377e-20  8.997593e-21  8.415035e-21  1.495130e-20
#> [291]  1.798059e-20  7.938294e-21 -1.004084e-20  1.086437e-20  1.227686e-20
#> [296]  1.325083e-20 -1.781339e-20  1.414904e-20  1.115756e-20 -1.902727e-20
#> [301]  1.993289e-20
#> 
#> $step
#> [1] 2
#> 
#> $plug.in
#> [1] 0.2411621
#> 
#> $support
#>   [1]  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [13] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [25] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [37] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [49] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [61] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [73] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [85] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [97] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [109] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [121] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [133] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [145] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [157] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [169] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [181] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [193] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [205] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [217] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [229] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [241] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [253] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [265] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [277] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [289] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [301] FALSE
#>