Inference for quadratic functional in high-dimensional linear regression model

Computes the bias-corrected estimator of the quadratic functional restricted to group G for the high-dimensional linear regression model and the corresponding standard error

QF(
  X,
  y,
  test.set,
  A = "sigma",
  init.Lasso = NULL,
  tau.vec = NULL,
  lambda = NULL,
  intercept = FALSE,
  mu = NULL,
  step = NULL,
  resol = 1.5,
  maxiter = 6
)

Arguments

X
Design matrix

y
Response variable

test.set
set of indices, G in the quadratic functional

A
Matrix A in the quadratic functional (either the population covariance matrix sigma or a known matrix of suitable dimension ; default = sigma)

init.Lasso
initial LASSO estimator for the regression vector (default = NULL)

tau.vec
Vector of enlargement factors for asymptotic variance of the bias-corrected estimator to handle super-efficiency (default = NULL)

lambda
Tuning parameter used in construction of initial LASSO estimator of the regression vector if init.Lasso = NULL (default = NULL)

intercept
Should intercept(s) be fitted (default = FALSE)

mu
Tuning parameter in construction of the projection direction (default = NULL)

step
Number of steps (< maxiter) to obtain the smallest mu that gives convergence of the optimization problem for constructing the projection direction (default = NULL)

resol
Resolution or the factor by which mu is increased/decreased to obtain the smallest mu that gives convergence of the optimization problem for constructing the projection direction (default = 1.5)

maxiter
Maximum number of steps along which mu is increased/decreased to obtain the smallest mu that gives convergence of the optimization problem for constructing the projection direction (default = 10)

Value

prop.est

The bias-corrected estimator of the quadratic functional restricted to G

sigma

Estimate of the error variance in the linear regression model

se

Standard error of the bias-corrected estimator

plug.in

Plug-in LASSO estimator for the quadratic functional restricted to G

References

Guo Z, Renaux C, B愼㸸uhlmann P, Cai TT (2019). “Group Inference in High Dimensions with Applications to Hierarchical Testing.” Unknown. https://statistics.rutgers.edu/home/zijguo/Guo-Renaux-Buhlman-Cai-Group-Inference.pdf.

Examples

X = matrix(sample(-2:2,100*400,replace = TRUE),nrow=100,ncol=400)
beta = (1:400)/25
y = X%*%beta + rnorm(100,0,1)
QF(X = X, y = y, test.set=c(30:50))

#> $prop.est
#>      [,1]
#> [1,]    0
#> 
#> $sigma
#> [1] 66321.45
#> 
#> $se
#> [1] 0.1
#> 
#> $plug.in
#> [1] 0
#>