Inference for quadratic functional in high-dimensional linear regression

Computes the bias-corrected estimator of \(\beta_G^{\top}\Sigma_{G,G}\beta_G\) for the following high-dimensional linear regression model $$y=X\beta+\epsilon$$ as $$\widehat{\beta_G^{\top}\Sigma_{G,G}\beta_G}=\widehat{\beta}_{G}^{\top} \widehat{\Sigma}_{G, G} \widehat{\beta}_{G}+\frac{2}{n} \widehat{u}^{\top} X^{\top}(y-X \widehat{\beta})$$ where \(\Sigma\) is the population design covariance matrix and \(G\subset{1,2,...,p}\)

QF_Cov(
  X,
  y,
  test.set,
  init.Lasso = NULL,
  tau.vec = NULL,
  lambda = NULL,
  intercept = FALSE,
  mu = NULL,
  step = NULL,
  resol = 1.5,
  maxiter = 6
)

Arguments

X
Design matrix, of dimensions nobs(n) x nvar(p)

y
Response variable

test.set
set of indices, G in \(\beta_G^{\top}\Sigma_{G,G}\beta_G\)

init.Lasso
initial LASSO estimator \(\widehat{\beta}\) (default = NULL)

tau.vec
Vector of enlargement factors for asymptotic variance of the estimator \(\widehat{\beta_G^{\top}\Sigma_{G,G}\beta_G}\) to handle super-efficiency (default = NULL)

lambda
Tuning parameter \(\lambda\) used in construction of initial LASSO estimator \(\widehat{\beta}\) if init.Lasso = NULL (default = NULL)

intercept
Should intercept(s) be fitted (default = FALSE)

mu
Tuning parameter \(\mu\) in construction of projection direction \(\widehat{u}\) (default = NULL)

step
Number of steps (< maxiter) to obtain the smallest mu that gives convergence of the optimization problem for constructing the projection direction \(\widehat{u}\) (default = NULL)

resol
Resolution or the factor by which mu is increased/decreased to obtain the smallest mu that gives convergence of the optimization problem for constructing the projection direction \(\widehat{u}\) (default = 1.5)

maxiter
Maximum number of steps along which mu is increased/decreased to obtain the smallest mu that gives convergence of the optimization problem for constructing the projection direction \(\widehat{u}\) (default = 6)

Value

prop.est

The bias-corrected estimator \(\widehat{\beta_G^{\top}\Sigma_{G,G}\beta_G}\)

sigma

Estimate of the variance of error term \(\epsilon\) in the linear regression model

se

Standard error of \(\widehat{\beta_G^{\top}\Sigma_{G,G}\beta_G}\)

plug.in

Plug-in LASSO estimator \(\widehat{\beta}_{G}^{\top} \widehat{\Sigma}_{G, G} \widehat{\beta}_{G}\)

References

Guo Z, Renaux C, B愼㸸uhlmann P, Cai TT (2019). “Group Inference in High Dimensions with Applications to Hierarchical Testing.” Unknown. https://statistics.rutgers.edu/home/zijguo/Guo-Renaux-Buhlman-Cai-Group-Inference.pdf.

Examples

X = matrix(sample(-2:2,100*400,replace = TRUE),nrow=100,ncol=400)
beta = (1:400)/25
y = X%*%beta + rnorm(100,0,1)
QF_Cov(X = X, y = y, test.set=c(30:50))

#> $prop.est
#>      [,1]
#> [1,]    0
#> 
#> $sigma
#> [1] 45561.04
#> 
#> $se
#> [1] 0
#> 
#> $plug.in
#> [1] 0
#>