Searches for the best step size and computes the projection direction in high dimensional linear regression

Searches for the best step size and computes the projection direction in high dimensional linear regression

Direction_searchtuning(X, loading, mu = NULL, resol = 1.5, maxiter = 10)

Arguments

X
Design matrix, of dimension \(n\) x \(p\)

loading
Loading, of length \(p\)

mu
Tuning parameter in construction of projection direction

resol
Resolution or the factor by which mu is increased/decreased to obtain the smallest mu that gives convergence of the optimization problem for constructing the projection direction (default = 1.5)

maxiter
Maximum number of steps along which mu is increased/decreased to obtain the smallest mu that gives convergence of the optimization problem for constructing the projection direction (default = 10)

Value

proj

The projection direction

Examples

n = 100
p = 400
X = matrix(sample(-2:2,n*p,replace = TRUE),nrow = n,ncol = p)
Direction_searchtuning(X,loading=c(1,rep(0,(p-1))))

#> [1] 3

#> $proj
#>   [1]  6.371366e-01  1.174962e-21 -4.827311e-22 -5.103170e-22  3.307916e-22
#>   [6]  1.092495e-22  1.815365e-21  1.531152e-21 -1.746862e-21 -1.345230e-21
#>  [11]  5.847732e-22 -5.535838e-22 -6.915321e-22  2.889573e-02  4.474369e-22
#>  [16]  7.473293e-22 -2.044716e-21 -4.198503e-22 -1.109705e-21 -6.576780e-22
#>  [21]  5.416451e-22  1.350079e-21 -5.010870e-22 -3.129621e-22 -7.348679e-23
#>  [26] -1.328753e-21 -5.261364e-22 -3.097145e-02  2.705607e-22 -9.219956e-22
#>  [31] -8.141399e-22 -1.500677e-21  7.696239e-22 -1.714290e-02  1.391946e-21
#>  [36]  5.282509e-02 -3.304213e-22  1.051629e-21  1.565843e-22 -5.334484e-23
#>  [41] -5.188493e-22  1.305276e-21 -1.382463e-21  5.581662e-22 -6.593759e-22
#>  [46]  5.143037e-23 -7.871113e-22  6.200100e-23 -1.093807e-23  1.436160e-21
#>  [51]  1.397262e-21  1.084036e-21  4.041964e-22 -1.534966e-21 -5.124085e-22
#>  [56]  8.676076e-22  7.884407e-02  4.773058e-22  2.538207e-22  3.126162e-22
#>  [61]  1.293039e-21 -6.569793e-22  8.035985e-23 -1.883375e-21  1.990154e-02
#>  [66]  1.188200e-02  1.911415e-21 -5.152477e-02  1.561029e-03  4.096357e-22
#>  [71]  1.994206e-21 -1.476547e-21 -3.698091e-22  9.051424e-23 -2.422595e-21
#>  [76] -1.938646e-21  1.924616e-21 -3.520909e-22 -1.709703e-21  1.856511e-02
#>  [81] -1.700678e-22 -5.124718e-22  1.886098e-22 -2.064681e-22 -4.055823e-22
#>  [86] -1.029464e-21  1.200385e-21 -3.583530e-22 -8.117260e-23  3.917399e-22
#>  [91]  7.292517e-22 -5.207249e-22 -1.526131e-22  3.154939e-22 -4.333163e-04
#>  [96]  2.088224e-21  5.311736e-22 -1.116486e-21  1.378710e-22 -8.947214e-22
#> [101]  2.365633e-22 -1.087367e-21 -1.340024e-21  8.688871e-22 -1.256005e-02
#> [106] -5.581448e-22  1.032271e-21  1.091787e-21 -1.796998e-21  4.067662e-22
#> [111] -2.300794e-21  2.218142e-21 -1.220561e-22  1.628536e-21 -9.762784e-02
#> [116] -1.299467e-22 -3.211211e-22 -9.634214e-22  7.898106e-22 -1.319937e-22
#> [121]  1.185536e-22 -8.760191e-22 -6.897404e-22 -1.607362e-21  1.798349e-21
#> [126] -9.543084e-22 -7.935359e-22  4.287221e-02 -6.343588e-22  1.180268e-21
#> [131] -1.732851e-22 -5.765454e-22  8.877242e-22 -3.546059e-22 -8.637660e-22
#> [136]  2.342496e-21  8.771924e-22  2.206426e-21  8.294090e-22  1.487714e-21
#> [141] -8.085651e-22 -1.107193e-21 -8.145254e-05  1.647095e-21 -4.934113e-22
#> [146]  4.516627e-22 -1.803469e-21 -7.839298e-22 -1.625959e-22 -4.495057e-23
#> [151]  8.533808e-22  3.679678e-22 -5.463913e-22  1.278696e-21  5.864482e-22
#> [156]  5.145986e-22 -9.331243e-23  1.367355e-22  8.463829e-22  4.585347e-22
#> [161]  4.524901e-22  4.627074e-22 -1.487717e-21  2.707531e-22  4.042551e-22
#> [166] -8.127293e-22 -3.914536e-22 -2.123479e-22  2.374561e-21  8.451483e-22
#> [171]  1.158692e-21 -1.315610e-21  1.189113e-21 -2.155072e-21  1.198479e-22
#> [176]  2.106855e-21  5.560976e-22 -4.531381e-22  9.666360e-22  2.914493e-22
#> [181]  4.396054e-22  7.139105e-22 -1.128320e-21  9.518967e-22  6.221198e-22
#> [186]  4.970480e-22 -4.385245e-22 -6.947925e-22  1.429812e-21  8.303807e-22
#> [191] -2.295416e-22  5.902455e-23 -2.895883e-22  3.772171e-22 -1.485670e-21
#> [196]  7.155886e-22 -9.354092e-22  9.638632e-22 -3.757535e-22  3.862733e-23
#> [201] -1.258763e-21  2.223593e-22 -4.458036e-22 -4.372761e-22  5.644577e-22
#> [206] -3.979387e-22  2.262625e-02 -1.155605e-21 -4.902679e-22  6.908849e-22
#> [211] -1.886993e-21 -4.503391e-22 -1.213747e-21  1.987065e-22 -6.830764e-22
#> [216] -1.482218e-22  8.560724e-22  4.006299e-22 -1.131448e-21 -3.760368e-22
#> [221]  6.454409e-22  1.293388e-22 -2.044119e-21 -9.879074e-22 -7.368280e-22
#> [226] -5.959669e-22 -1.139089e-21  1.407140e-21 -5.837991e-22 -2.491273e-22
#> [231] -8.986387e-22  8.379960e-22 -2.069463e-22  1.314205e-21 -3.558843e-22
#> [236]  5.480301e-23  2.740925e-22  4.089459e-02 -2.293998e-22 -1.235693e-21
#> [241] -6.471500e-22 -8.803264e-22  1.547346e-22 -1.322917e-21 -1.313473e-22
#> [246] -1.361357e-02 -8.827443e-22  3.257612e-22 -6.241428e-22  1.615594e-21
#> [251]  3.862572e-03 -4.140259e-22  3.258855e-22 -6.532532e-22 -5.376938e-22
#> [256] -1.564877e-21  2.581912e-22 -2.966148e-21 -9.436761e-22  8.608369e-22
#> [261]  9.234634e-22  1.437168e-21  1.373578e-21 -1.538833e-21  1.221245e-02
#> [266]  4.316122e-22 -3.443113e-22 -1.379341e-21 -1.344744e-21 -5.161602e-22
#> [271]  5.906446e-22 -1.210193e-21 -6.703068e-22 -5.209920e-02 -8.949413e-23
#> [276]  1.158133e-21  5.094671e-23 -1.299928e-21  6.473282e-22 -1.527133e-22
#> [281]  1.711620e-02  2.766328e-02  3.645736e-02  4.155585e-22  5.511373e-22
#> [286] -1.184496e-21  5.290720e-22 -9.047861e-22  1.941758e-21 -8.525463e-22
#> [291]  1.696760e-22  8.957082e-22  1.698368e-21  1.735448e-21  5.493357e-22
#> [296]  4.645564e-22 -1.159391e-21 -1.517447e-21 -9.558499e-22 -5.075226e-22
#> [301]  8.128766e-22 -6.517545e-22 -2.941439e-22 -7.041529e-23  1.149029e-21
#> [306] -4.351108e-23  1.983185e-22 -3.815949e-22 -5.915094e-22  4.280568e-22
#> [311] -5.628486e-22  4.801385e-22 -8.064509e-22  2.246017e-22 -1.743385e-21
#> [316]  1.492001e-22 -1.893400e-02  4.630747e-22  2.051751e-21  1.793995e-21
#> [321] -1.063547e-21 -9.970808e-22 -1.176070e-21 -1.435574e-21 -8.999900e-22
#> [326] -1.111004e-21  8.639129e-22 -1.251934e-21  1.829580e-22 -3.705345e-22
#> [331]  4.143814e-22  1.314982e-21  1.855534e-21  8.390906e-22  4.867647e-23
#> [336]  2.838627e-22 -9.216714e-22 -1.650048e-21  7.876563e-22  8.385083e-22
#> [341]  1.948141e-21 -2.461212e-22 -1.700320e-21 -1.043022e-22  1.014710e-21
#> [346] -8.876173e-22  2.014191e-21  2.373673e-21 -1.637069e-03  1.278823e-21
#> [351] -1.627874e-21  6.929576e-02 -3.715799e-22 -6.352344e-22  2.073057e-22
#> [356] -6.963256e-22  1.269555e-21 -1.273562e-21  8.626909e-23  5.429118e-22
#> [361] -7.593645e-22  1.974708e-22  2.092305e-21  1.997968e-21  8.243520e-22
#> [366]  4.855463e-23 -2.618762e-22  3.637512e-02  1.324231e-21  3.587650e-22
#> [371] -9.998108e-22  9.187236e-22 -1.283417e-21 -2.527937e-21 -1.411560e-22
#> [376] -1.707832e-21 -4.456961e-22 -1.121449e-21 -2.647166e-02  3.843692e-22
#> [381] -5.741991e-22  9.457363e-22 -8.296956e-02 -1.427740e-21  5.599736e-22
#> [386]  1.469752e-21 -7.612525e-22  4.282217e-22  4.633870e-22 -3.560459e-22
#> [391] -8.431316e-22 -8.501184e-22 -1.013098e-21 -4.437824e-22 -1.770161e-22
#> [396]  7.955977e-22  1.727041e-21 -7.019979e-22 -2.377583e-22  5.353252e-22
#> 
#> $step
#> [1] 3
#>