Inference for linear functional in the high dimensional linear regression model

Computes the bias corrected estimator of linear functional loading\(^{\top}\beta\) for the high dimensional linear regression model \(Y_i=X_i^{\top}\beta + \epsilon_i\) and the corresponding standard error.

LF(
  X,
  y,
  loading,
  init.Lasso = NULL,
  lambda = NULL,
  intercept = TRUE,
  mu = NULL,
  step = NULL,
  resol = 1.5,
  maxiter = 10
)

Arguments

X
Design matrix, of dimension \(n\) x \(p\)

y
Outcome vector, of length \(n\)

loading
Loading, of length \(p\)

init.Lasso
Initial LASSO estimator of the regression vector (default = NULL)

lambda
The tuning parameter in construction of LASSO estimator of the regression vector (default = NULL)

intercept
Should intercept(s) be fitted (default = TRUE)

mu
The tuning parameter in construction of projection direction (default = NULL)

step
Number of steps (< maxiter) to obtain the smallest mu that gives convergence of the optimization problem for constructing the projection direction (default = NULL)

resol
Resolution or the factor by which mu is increased/decreased to obtain the smallest mu that gives convergence of the optimization problem for constructing the projection direction (default = 1.5)

maxiter
Maximum number of steps along which mu is increased/decreased to obtain the smallest mu that gives convergence of the optimization problem for constructing the projection direction (default = 10)

Value

prop.est

The bias-corrected estimator for the linear functional

sigma

Estimate of the error variance in the linear regression model

se

Standard error of the bias-corrected estimator

proj

Optimal projection direction

step

Number of steps (< maxiter) to obtain the smallest mu that gives convergence of the optimization problem for constructing the projection direction

plug.in

Plug-in LASSO estimator for the linear functional

References

Cai T, Cai TT, Guo Z (2019). “Individualized Treatment Selection: An Optimal Hypothesis Testing Approach In High-dimensional Models.” Unknown. https://statistics.rutgers.edu/home/zijguo/Cai-Cai-Guo-ITE-Selection.pdf.

Examples

n = 100
p = 400
X = matrix(sample(-2:2,n*p,replace = TRUE),nrow = n,ncol = p)
beta = (1:p)/25
y = X%*%beta + rnorm(n,0,1)
LF(X = X, y = y, loading = c(1,rep(0,(p-1))), intercept = TRUE)

#> [1] 3
#> [1] "step is 3"

#> $prop.est
#>          [,1]
#> [1,] 28.66721
#> 
#> $sigma
#> [1] 75324.4
#> 
#> $se
#> [1] 8174.09
#> 
#> $proj
#>   [1]  1.257514e+00  8.572327e-01 -6.062040e-21 -4.397018e-20  2.253397e-20
#>   [6]  1.515383e-20  3.916893e-20  3.493509e-20 -6.439319e-20  1.271803e-20
#>  [11] -1.024120e-01 -3.272223e-20  4.144288e-03  2.005773e-21  3.774126e-20
#>  [16]  3.720129e-20  1.084970e-20 -6.140566e-20  5.787387e-20  3.360670e-20
#>  [21]  3.396410e-02 -2.624619e-02  6.125256e-20  2.705553e-02  5.670542e-20
#>  [26] -2.085271e-20  1.483731e-20  4.213273e-20  1.117504e-20  1.344371e-20
#>  [31] -2.215989e-02  4.784489e-20  7.286528e-20  8.013656e-20  7.226296e-21
#>  [36]  3.187829e-20 -6.301693e-20 -1.761147e-20  3.373469e-20 -1.182779e-02
#>  [41] -8.243996e-21  1.732332e-21 -6.065341e-20 -1.311560e-01 -1.832626e-22
#>  [46] -4.850193e-20  1.192091e-19 -1.947964e-21 -1.386226e-01 -1.714244e-20
#>  [51] -8.316294e-20  3.479177e-20  3.325343e-20  3.210858e-20 -4.342920e-20
#>  [56] -3.850690e-20  7.889542e-20  1.472494e-01  6.700434e-20 -2.696153e-20
#>  [61] -1.003295e-20 -7.557089e-20  2.753871e-20  6.523264e-20 -1.199700e-20
#>  [66] -6.331675e-21  8.065639e-20  5.333080e-20 -1.097940e-19 -2.105816e-20
#>  [71]  9.213319e-20  2.767505e-20  9.472143e-21  7.960875e-20 -6.066464e-20
#>  [76] -9.485528e-20 -8.890334e-21  3.357035e-21  5.239715e-20 -2.422013e-20
#>  [81]  8.320560e-20  3.944175e-20  1.706101e-02  6.885144e-20 -1.960882e-21
#>  [86] -2.226545e-20 -6.785976e-02  4.411855e-20  3.643847e-02 -7.601880e-21
#>  [91] -3.550360e-20 -2.854106e-20 -6.362096e-03  5.479566e-20 -1.938124e-20
#>  [96]  4.818643e-20 -8.663086e-21  1.203154e-19 -5.077516e-02 -9.699804e-21
#> [101] -4.612805e-20 -7.565840e-20  6.252380e-20 -1.777149e-22 -3.672754e-20
#> [106]  3.170859e-21 -8.774295e-02 -5.861032e-20  1.334222e-20  6.434803e-21
#> [111] -8.964238e-22 -5.725451e-21  7.273882e-03 -4.288110e-20  1.511329e-01
#> [116]  4.165123e-20  8.117184e-20  2.572044e-20 -7.064629e-20 -8.676282e-20
#> [121]  8.425579e-20 -2.112234e-20 -6.539306e-20 -1.929549e-20 -8.684114e-21
#> [126] -4.131067e-20 -1.246936e-20  3.844512e-20  3.410877e-20 -8.074944e-21
#> [131] -1.204269e-01 -3.977429e-02  6.700902e-20 -2.404493e-20 -6.839906e-20
#> [136]  7.065924e-21  3.644595e-20 -5.270242e-20  6.398977e-20 -1.490061e-20
#> [141]  4.322372e-20 -2.835122e-21  7.488654e-20  1.015908e-01  4.523859e-20
#> [146] -5.048052e-20 -2.424582e-20  4.895005e-21  1.664007e-20  6.443840e-20
#> [151] -8.751690e-20  1.703806e-01 -6.434661e-20  7.898485e-21 -7.093267e-20
#> [156] -2.467604e-20 -2.664932e-02 -2.597864e-20 -4.193754e-20 -4.931852e-02
#> [161]  3.595192e-20  2.272305e-20  3.611118e-02 -3.906762e-20  3.584521e-20
#> [166]  3.016982e-20 -1.132956e-20  2.684337e-02  1.489092e-04 -3.135654e-20
#> [171] -3.132764e-20  8.933126e-02 -4.771816e-20 -5.125727e-20  1.650834e-21
#> [176] -7.141752e-20  8.714346e-21  2.555913e-20 -3.466517e-20  4.199661e-20
#> [181] -1.155353e-20 -1.968831e-20 -2.465658e-20 -2.731303e-21 -2.821280e-20
#> [186] -3.197396e-02 -6.517509e-20 -5.241804e-20  4.353812e-20 -2.883751e-20
#> [191] -3.250256e-20  5.646724e-20  2.518503e-21  4.586433e-20 -2.673907e-20
#> [196]  5.742103e-02 -1.860149e-01  1.819932e-20 -1.304126e-20  4.571182e-20
#> [201] -2.561222e-20  1.764897e-20 -5.325604e-20  6.796045e-21  1.723706e-20
#> [206]  5.643811e-20 -9.935590e-02 -6.350202e-02 -4.388489e-02  1.018663e-19
#> [211]  2.924693e-20 -8.914481e-20  7.421654e-21 -1.259620e-19  2.315729e-20
#> [216] -4.703376e-20  2.190623e-21  5.736902e-21 -4.643924e-21  4.010273e-20
#> [221] -3.533190e-20 -8.695122e-21 -5.273572e-21  2.137014e-20  2.671415e-20
#> [226]  1.769766e-20  5.119627e-20  4.028671e-20 -7.635353e-21  3.067412e-22
#> [231] -1.574524e-20 -1.567844e-01  3.389991e-20 -2.190402e-20 -3.047688e-20
#> [236] -1.566163e-21 -4.685205e-02 -4.599473e-02  1.037670e-20  3.200604e-20
#> [241] -9.570720e-21 -7.235684e-20 -1.812611e-20 -1.689270e-20  7.632414e-02
#> [246]  6.137970e-20  1.780033e-20 -3.274704e-20 -2.581767e-20 -1.601701e-20
#> [251]  3.205113e-20  3.668089e-20  2.946003e-20  3.005552e-20 -2.073360e-01
#> [256] -6.555488e-02  4.223927e-20 -3.132402e-20 -6.854129e-21  4.079931e-20
#> [261] -2.760276e-20  1.586300e-20 -9.571867e-20 -7.320468e-02 -2.404418e-20
#> [266]  5.841202e-21 -2.084677e-20 -6.551879e-20  5.957364e-20  7.240324e-20
#> [271] -4.566765e-02 -5.979999e-02  2.016318e-20  3.473971e-02  3.867008e-20
#> [276]  1.591053e-20 -9.084430e-21 -7.489840e-20 -3.473257e-20  3.381374e-20
#> [281] -2.196354e-20 -6.282455e-21 -3.570850e-20 -5.939226e-03  1.052089e-21
#> [286]  3.180050e-20 -1.501485e-20 -1.542199e-20 -4.106321e-02 -1.511121e-20
#> [291] -1.081461e-20 -7.718559e-02 -8.185925e-20  3.154422e-20  5.222460e-20
#> [296]  4.756506e-20  5.341337e-20  4.332347e-20 -7.317070e-20  7.521798e-20
#> [301] -3.045831e-20  1.795563e-20 -4.762145e-20  8.053944e-21 -4.494019e-21
#> [306]  3.626217e-20 -1.918636e-02  3.254278e-21  1.904393e-20  6.585435e-02
#> [311] -1.707344e-20 -1.804804e-01 -5.287030e-20  8.458950e-20 -3.273827e-20
#> [316] -3.134034e-20 -4.738240e-20  1.644237e-01 -4.867880e-20 -2.764741e-20
#> [321] -6.014539e-21 -8.492960e-20 -1.129279e-01 -8.821141e-20  7.975917e-21
#> [326]  3.265983e-20  1.594889e-20  7.082548e-20 -4.272301e-21 -1.000845e-19
#> [331]  3.998027e-20  3.353114e-20  2.541499e-20 -1.973303e-20  2.977675e-20
#> [336]  1.922209e-20 -8.198586e-02  9.105209e-20 -3.494830e-20 -1.773002e-02
#> [341]  2.423550e-20  4.365715e-20  2.628392e-02 -2.641706e-20 -1.134876e-19
#> [346] -5.541370e-20 -5.139464e-20 -2.950516e-21  5.646237e-21 -7.387875e-20
#> [351] -2.588246e-02  1.207197e-19 -5.892585e-20 -2.562438e-20 -5.040859e-21
#> [356]  2.457290e-02  4.053930e-21 -5.219017e-20 -4.238400e-21  8.789450e-20
#> [361]  5.233576e-20  4.588396e-20 -2.051324e-20  9.630875e-20 -4.465754e-20
#> [366] -2.610809e-20  2.546674e-20  9.172254e-02 -5.067241e-20 -3.885545e-20
#> [371] -2.698282e-02  6.157517e-20 -3.693096e-20  1.857222e-20 -6.002559e-20
#> [376] -1.117626e-20  1.318620e-20 -4.962952e-20  1.061803e-20 -1.208747e-20
#> [381] -5.848367e-20  4.508708e-02 -3.218423e-20 -6.766345e-21  5.635350e-20
#> [386]  4.601161e-21  5.295889e-02 -7.012872e-21  7.637920e-02 -2.255582e-02
#> [391] -5.905588e-20 -9.546527e-03  2.138393e-20  6.210981e-20  7.076055e-20
#> [396] -1.638477e-20  1.078277e-20  7.062319e-02  3.590730e-20  1.374276e-20
#> [401]  5.084093e-20
#> 
#> $step
#> [1] 3
#> 
#> $plug.in
#> [1] 26.06835
#>