Inference for linear functional in the high-dimensional logistic regression model

Computes the bias corrected estimator of the linear functional for the high-dimensional logistic regression model and the corresponding standard error.

LF_logistic(
  X,
  y,
  loading,
  weight,
  lambda = NULL,
  intercept = FALSE,
  mu = NULL,
  step = NULL,
  resol = 1.5,
  maxiter = 6
)

Arguments

X
Design matrix

y
Response vector

loading
observation vector in the linear functional

weight
The vector of weights used in correcting the plug-in estimator

lambda
Tuning parameter in construction of the LASSO estimator of the regression vector (default = NULL)

intercept
Should intercept(s) be fitted (default = FALSE)

mu
Tuning parameter in construction of the projection direction (default = NULL)

step
Number of steps (< maxiter) to obtain the smallest mu that gives convergence of the optimization problem for constructing the projection direction (default = NULL)

resol
Resolution or the factor by which mu is increased/decreased to obtain the smallest mu that gives convergence of the optimization problem for constructing the projection direction (default = 1.5)

maxiter
Maximum number of steps along which mu is increased/decreased to obtain the smallest mu that gives convergence of the optimization problem for constructing the projection direction (default = 6)

Value

prop.est

The bias corrected estimator of the linear functional

se

Standard error of the bias-corrected estimator

proj

Optimal projection direction

step

Number of steps (< maxiter) to obtain the smallest mu that gives convergence of the optimization problem for constructing the projection direction

plug.in

Plug-in LASSO estimator of the linear functional

support

Estimated support for the regression vector (using LASSO estimate)

References

Guo Z, Rakshit P, Herman DS, Chen J (2019). “Inference for Case Probability in High-dimensional Logistic Regression.” Unknown.

Examples

LF_logistic(X = matrix(sample(-2:2,50*300,replace = TRUE),nrow=50,ncol=300),
                      y = rbinom(50,1,0.5), loading = c(1,rep(0,299)),
                      intercept = TRUE, weight = rep(1,50))

#> [1] 2
#> [1] "step is 2"

#> $prop.est
#> [1] -1.038514
#> 
#> $se
#> [1] 0.4335102
#> 
#> $proj
#>   [1]  5.574850e+00  4.848058e+00 -3.599076e-18 -1.363495e-18 -1.345742e-18
#>   [6]  3.824917e-18 -7.442853e-19 -5.977451e-19 -1.704153e-18 -5.696649e-19
#>  [11]  4.176139e-18  4.467138e-18  5.709432e-18 -7.606455e-20  1.222008e-18
#>  [16] -6.402554e-19  1.428725e-20  3.526555e-18  1.806973e-19  3.083369e-20
#>  [21]  4.160646e-19 -8.621136e-18 -1.962789e-18 -2.164247e-18 -4.912778e-18
#>  [26] -2.984777e-18 -9.370494e-19 -3.919139e-18 -9.090991e-19  4.583261e-18
#>  [31]  2.840593e-18 -1.988693e-18  1.199757e-18 -4.160560e-18  4.095384e-18
#>  [36] -6.021867e-19  1.679977e-18  5.034773e-21  1.686633e-19  5.467321e-19
#>  [41]  1.546475e-18  2.295762e-18  4.905983e-19  8.591140e-19  2.921238e-18
#>  [46]  1.981508e-18 -7.233344e-19 -4.086012e-18 -2.876232e-18 -3.749395e-18
#>  [51] -6.651166e-19 -9.522972e-19  3.936506e-18  6.869373e-18 -1.320083e-18
#>  [56] -1.902107e-18  1.951563e-18  2.888814e-18 -4.235576e-18  2.641502e-18
#>  [61]  1.545657e-18  1.221678e-18  3.128040e-18  8.220304e-20 -1.202863e-18
#>  [66] -1.228998e-18 -2.761584e-18 -2.215954e-20  2.370658e-18  4.395860e-19
#>  [71]  8.467144e-19 -4.396460e-18 -6.717586e-19 -1.739047e-18 -7.695882e-18
#>  [76] -3.818881e-18  1.937993e-18  2.159475e-18 -8.060104e-19  2.060804e-18
#>  [81]  6.547879e-19 -1.590037e-18 -3.139190e-18  8.447202e-18  2.908480e-18
#>  [86] -9.064827e-19 -8.194731e-20 -1.566265e-18  1.837557e-18 -2.383062e-18
#>  [91] -3.085878e-01 -3.107663e-18 -5.225923e-18  4.162778e-18 -3.766556e-18
#>  [96]  1.971358e-18  1.351415e-18 -2.391059e-18 -2.772000e-18  2.640272e-18
#> [101] -6.275131e-19  2.140662e-18 -1.676536e-18  1.696829e-18 -1.051464e-18
#> [106] -8.769331e-19 -1.941077e-18  2.357152e-18  2.705482e-18  3.001299e-18
#> [111]  4.135180e-18 -7.432610e-18 -2.595965e-18  4.192823e-18 -1.232632e-18
#> [116]  5.976208e-19  6.599978e-19 -4.632439e-18 -1.970630e-18  1.236024e-18
#> [121]  3.126534e-18  1.909582e-18 -4.223333e-19 -1.121853e-18 -4.919902e-18
#> [126] -1.551466e-20 -2.854923e-19 -3.417207e-18 -4.875550e-18  1.281095e-18
#> [131]  1.976817e-18  1.617658e-19 -7.274567e-19  4.796476e-18  4.292325e-18
#> [136] -6.941040e-19 -1.309757e-20  2.091553e-21 -1.042080e-18  7.112121e-19
#> [141] -2.584621e-18 -7.967828e-19 -5.932856e-18 -3.504578e-18 -3.701482e-18
#> [146] -6.889338e-20 -2.538011e-19 -5.250905e-18  2.359103e-18  7.547523e-19
#> [151]  2.057769e-18  4.756141e-01  5.686918e-18  1.966425e-18  2.770684e-18
#> [156]  3.055827e-18  2.525348e-18 -1.239517e-18  1.805897e-18 -4.930301e-19
#> [161]  3.197354e-18 -3.083437e-18 -3.293118e-18  1.035115e-18  4.227791e-18
#> [166] -5.725930e-19  2.923584e-18  7.891435e-19  3.093258e-18  2.565756e-18
#> [171]  1.133567e-18 -1.254554e-18 -4.909027e-18 -5.562957e-18  9.590556e-19
#> [176]  3.081956e-18  7.292343e-18  7.211419e-18 -8.186792e-19 -4.938134e-19
#> [181]  1.853628e-19 -4.191571e-19  4.738663e-18  5.322937e-19  2.400756e-18
#> [186] -2.324509e-18  1.923611e-18 -7.692877e-19 -3.316528e-18  2.891863e-18
#> [191] -1.936698e-18 -4.037067e-18 -6.519442e-18  1.508147e-18 -3.156605e-18
#> [196]  1.486927e-18  7.004556e-19 -7.690414e-18  3.236163e-18  4.105784e-18
#> [201] -2.859199e-18  2.323941e-18  1.618471e-18 -1.156852e-18  1.311248e-18
#> [206]  3.587221e-19  2.204670e-19 -1.104487e-18  2.542467e-18 -2.724110e-18
#> [211] -2.938952e-18  1.222815e-18 -4.807733e-18 -3.025483e-18 -3.918087e-18
#> [216]  4.778316e-18  1.501195e-18  3.553235e-18  7.331299e-19 -1.657216e-18
#> [221] -1.317605e-19  1.874935e-18  3.471980e-18  5.441234e-19  1.707391e-18
#> [226]  5.659143e-18  2.854537e-18  3.907969e-18 -3.393581e-18  1.060425e-18
#> [231]  1.682509e-18 -8.400589e-19 -2.086418e-19 -1.687018e-18 -5.090137e-18
#> [236]  3.088055e-19 -4.207414e-18  2.592572e-18 -1.599516e-18 -1.897408e-18
#> [241] -1.166569e-18  5.838478e-19 -3.018531e-18 -3.069918e-18 -3.666946e-18
#> [246]  1.839551e-18  2.942035e-18  1.401964e-18 -4.752159e-18  9.655737e-20
#> [251]  7.403506e-19  3.771311e-18 -1.298629e-18  1.099691e-18 -5.357472e-19
#> [256] -1.452347e-18  3.267920e-18  1.905194e-18  4.135890e-18 -5.861262e-19
#> [261]  4.230694e-18  6.310376e-19 -1.371660e-18  9.162071e-19 -3.089776e-19
#> [266]  3.790124e-19 -2.654183e-18  1.056014e-18  2.338317e-19  1.412413e-18
#> [271]  3.675455e-18  2.796995e-18  2.659549e-18 -2.727636e-18 -5.511373e-18
#> [276] -2.837335e-18 -2.379737e-18 -2.455984e-19  1.965706e-18  2.036268e-18
#> [281]  2.651374e-19 -4.019424e-18  1.120925e-18  7.433167e-19 -1.821631e-18
#> [286]  1.475128e-18 -1.627573e-18  6.428882e-18  1.301624e-18  1.702428e-18
#> [291] -8.170679e-19 -2.193802e-18 -7.046941e-18  2.002334e-19 -1.880122e-18
#> [296] -4.054061e-18  3.221037e-18  6.832947e-19  1.384159e-18  2.288203e-18
#> [301]  3.349258e-18
#> 
#> $step
#> [1] 2
#> 
#> $plug.in
#> [1] -0.8780846
#> 
#> $support
#>   [1]  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [13] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [25] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [37] FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE
#>  [49] FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE
#>  [61] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [73] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE
#>  [85] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE
#>  [97] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE
#> [109] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE
#> [121] FALSE FALSE FALSE  TRUE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE
#> [133] FALSE FALSE  TRUE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE
#> [145] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [157] FALSE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [169] FALSE FALSE FALSE FALSE  TRUE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE
#> [181] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [193] FALSE FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [205] FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE  TRUE FALSE FALSE FALSE
#> [217] FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE
#> [229] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [241] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE  TRUE
#> [253] FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [265] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [277] FALSE FALSE FALSE FALSE  TRUE FALSE FALSE  TRUE FALSE FALSE FALSE FALSE
#> [289] FALSE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [301] FALSE
#>