Inference for linear functional in the high-dimensional logistic regression model

Computes the bias corrected estimator of the linear functional for the high-dimensional logistic regression model and the corresponding standard error.

LF_logistic(
  X,
  y,
  loading,
  weight,
  init.Lasso = NULL,
  intercept = TRUE,
  mu = NULL,
  step = NULL,
  resol = 1.5,
  maxiter = 6
)

Arguments

X
Design matrix, of dimension nvar(\(n\))xnobs(\(p\))

y
Response vector, of length \(n\)

loading
observation vector in the linear functional, of length \(p\)

weight
The vector, of length \(n\), of weights used in correcting the plug-in estimator

init.Lasso
initial LASSO estimator of the regression vector (default = NULL)

intercept
Should intercept(s) be fitted (default = TRUE)

mu
Tuning parameter in construction of the projection direction (default = NULL)

step
Number of steps (< maxiter) to obtain the smallest mu that gives convergence of the optimization problem for constructing the projection direction (default = NULL)

resol
Resolution or the factor by which mu is increased/decreased to obtain the smallest mu that gives convergence of the optimization problem for constructing the projection direction (default = 1.5)

maxiter
Maximum number of steps along which mu is increased/decreased to obtain the smallest mu that gives convergence of the optimization problem for constructing the projection direction (default = 6)

Value

prop.est

The bias corrected estimator of the linear functional

se

Standard error of the bias-corrected estimator

proj

Optimal projection direction

step

Number of steps (< maxiter) to obtain the smallest mu that gives convergence of the optimization problem for constructing the projection direction

plug.in

Plug-in LASSO estimator of the linear functional

support

Estimated support for the regression vector (using LASSO estimate)

References

Guo Z, Rakshit P, Herman DS, Chen J (2019). “Inference for Case Probability in High-dimensional Logistic Regression.” Unknown.

Examples

LF_logistic(X = matrix(sample(-2:2,50*300,replace = TRUE),nrow=50,ncol=300),
                      y = rbinom(50,1,0.5), loading = c(1,rep(0,299)),
                      intercept = TRUE, weight = rep(1,50))

#> Warning: longer object length is not a multiple of shorter object length

#> [1] 2
#> [1] "step is 2"

#> $prop.est
#> [1] 0.171091
#> 
#> $se
#> [1] 0.2386363
#> 
#> $proj
#>   [1]  1.705200e+00  1.443144e+00  1.049184e-20 -2.279830e-20 -1.459109e-20
#>   [6]  1.614109e-20  2.345555e-20 -2.806371e-20 -1.248395e-20 -1.994544e-20
#>  [11] -1.986145e-20  4.000462e-20  2.012333e-20 -3.633683e-20  2.894078e-21
#>  [16]  3.865969e-21 -1.515960e-20 -1.407974e-21 -1.735501e-20 -3.884473e-20
#>  [21] -3.807159e-20 -2.298529e-20 -2.623879e-20  4.384593e-20  9.623911e-21
#>  [26]  1.960222e-20  1.027197e-20 -1.591352e-20 -6.637170e-21 -1.280178e-20
#>  [31]  3.547712e-20 -1.761525e-20 -2.413398e-20  7.929610e-22 -6.918784e-20
#>  [36]  1.997197e-21 -3.474912e-20  4.074902e-21 -1.917044e-20  2.771487e-21
#>  [41]  6.910004e-21  8.961425e-22  2.863480e-21 -5.543765e-20 -3.809566e-20
#>  [46]  2.783315e-20 -1.670178e-20 -1.009258e-21  2.546416e-21 -1.057530e-20
#>  [51] -9.743768e-21 -2.047550e-20  2.782709e-21 -2.473229e-21  1.072781e-21
#>  [56] -9.738598e-21 -6.591771e-20 -4.616533e-20  4.285347e-20  3.210983e-20
#>  [61] -6.009165e-21 -1.473382e-20  1.130462e-20  2.020351e-20  2.123668e-20
#>  [66] -3.944253e-21  5.491160e-21 -1.699559e-20  1.502382e-21 -3.869810e-21
#>  [71]  3.278909e-20 -1.497345e-23 -3.291842e-20  2.309926e-20 -4.308789e-21
#>  [76]  4.398682e-21  3.162805e-20  3.019884e-21  4.290231e-20  6.870625e-21
#>  [81]  5.060333e-20  1.272463e-20  1.890467e-20 -4.488100e-21  1.037720e-20
#>  [86] -2.617738e-20  8.170859e-21  4.054546e-21  3.557556e-20  3.635668e-20
#>  [91] -8.791695e-21  4.895599e-20 -1.322209e-20 -2.010957e-20  1.500060e-20
#>  [96]  3.553478e-20  2.711597e-20 -1.641077e-20 -2.718856e-21  3.183844e-03
#> [101]  1.363597e-20  2.840228e-20  5.212118e-20  1.579081e-20 -1.795609e-20
#> [106]  7.385842e-20 -1.206527e-20  2.978791e-20  2.390269e-20  3.065427e-20
#> [111]  2.588829e-20 -1.630812e-20  9.449231e-21  4.337521e-21 -1.790502e-20
#> [116] -6.420261e-21  4.000389e-20  1.584141e-20  3.857152e-20  8.249978e-21
#> [121] -1.105772e-20 -3.082510e-20 -2.196072e-20 -4.630964e-20 -1.761215e-20
#> [126] -1.308499e-20  2.419614e-20 -1.480438e-20  1.397644e-20  5.578898e-20
#> [131] -2.618157e-20 -3.206351e-20 -4.692248e-20 -3.394858e-20  2.933718e-20
#> [136]  2.733332e-20  3.071978e-20 -1.160565e-20 -8.744836e-21  2.135204e-20
#> [141] -4.046378e-20 -9.249227e-21 -7.007925e-21  2.452332e-20 -1.008222e-20
#> [146] -3.474735e-21 -1.410995e-20 -4.660676e-22 -5.892429e-21 -1.480628e-20
#> [151]  1.173786e-20 -2.961378e-20  5.345104e-21  1.387197e-20  1.776999e-20
#> [156]  8.462861e-21 -1.379208e-20 -1.789920e-20 -1.539944e-20 -3.660619e-20
#> [161]  1.285391e-20 -2.418115e-21  3.808894e-20 -6.124489e-21 -6.940745e-21
#> [166] -1.480533e-20 -1.728659e-20  5.769612e-21  1.262033e-21  2.921381e-22
#> [171] -1.216306e-20  7.122412e-20  2.675011e-20  2.377731e-20  5.340335e-21
#> [176] -3.607785e-21  1.242245e-20  1.180237e-20 -8.774878e-21  4.232887e-20
#> [181]  2.535407e-21  2.918879e-20 -7.975972e-21  6.502885e-21  1.287273e-20
#> [186]  1.338492e-20 -4.926845e-20 -1.681885e-20 -5.255058e-23 -3.418944e-21
#> [191]  1.307916e-20  9.850303e-21 -3.599312e-21  4.033974e-21  2.373026e-20
#> [196] -2.099064e-20  2.049058e-20 -2.113956e-20 -2.853982e-20 -9.996077e-21
#> [201] -1.427279e-21 -1.274802e-21 -1.080326e-20  2.682637e-21  1.660100e-21
#> [206] -8.599254e-21 -9.989906e-21 -4.648292e-20 -3.887982e-21  2.467686e-22
#> [211]  2.475230e-20 -4.856135e-20 -2.539844e-20 -5.019550e-20 -2.165517e-20
#> [216]  7.868107e-21  2.724663e-21 -1.101409e-20 -3.237965e-20  1.129671e-20
#> [221] -1.069191e-20  1.943962e-20 -9.392313e-22  1.719347e-20  2.925563e-20
#> [226] -2.272046e-20 -1.234013e-20 -5.965006e-21 -2.088057e-20 -1.202692e-20
#> [231]  1.470995e-20 -2.018243e-21 -3.395532e-20  7.534612e-21  6.281386e-21
#> [236] -3.613716e-20  1.653295e-20 -4.927842e-21 -4.285864e-21  1.067559e-20
#> [241] -1.494116e-20  2.070795e-01  5.504316e-21  5.631069e-21 -8.408501e-21
#> [246]  1.615167e-20  3.862994e-20 -3.456178e-20 -2.962223e-21 -2.913560e-20
#> [251] -1.025099e-20  4.373900e-21 -3.546768e-20  2.741879e-20  1.524938e-20
#> [256]  5.793529e-22  5.298774e-20 -2.157422e-20  9.810359e-21  3.717226e-20
#> [261]  3.049779e-20  5.761590e-21 -1.996171e-23 -1.018558e-20 -1.602781e-20
#> [266] -2.356311e-20  1.282516e-20 -2.499405e-21 -3.835366e-20 -1.004853e-20
#> [271]  4.049105e-20  4.505365e-21  1.231034e-20  1.761146e-20  1.720894e-20
#> [276] -4.783428e-21 -3.494966e-20  1.875562e-20 -2.054449e-20  3.438707e-20
#> [281] -3.309454e-20  2.171383e-20  1.152301e-20 -4.457368e-20 -3.761527e-20
#> [286] -6.730135e-22 -1.170121e-20  8.569427e-21  8.045115e-21  1.432911e-20
#> [291]  1.719639e-20  7.603364e-21 -9.592684e-21  1.040036e-20  1.179287e-20
#> [296]  1.271101e-20 -1.700730e-20  1.355715e-20  1.068672e-20 -1.819678e-20
#> [301]  1.902868e-20
#> 
#> $step
#> [1] 2
#> 
#> $plug.in
#> [1] 0
#> 
#> $support
#> NULL
#>