Inference for linear functional in the high dimensional logistic regression model

Computes the bias corrected estimator of the linear functional loading\(^{\top}\beta\) for the high dimensional logistic regression model \(Y_i|X_i \sim \) Bernoulli\((\frac{e^{X_i^{\top}\beta}}{1+e^{X_i^{\top}\beta}})\) and the corresponding standard error.

LF_logistic(
  X,
  y,
  loading,
  weight,
  intercept = TRUE,
  init.Lasso = NULL,
  mu = NULL,
  step = NULL,
  resol = 1.5,
  maxiter = 10
)

Arguments

X
Design matrix, of dimension \(n\) x \(p\)

y
Outcome vector, of length \(n\)

loading
Loading, of length \(p\)

weight
The weight vector, of length \(n\), used in correcting the plug-in estimator

intercept
Should intercept(s) be fitted (default = TRUE)

init.Lasso
Initial LASSO estimator of the regression vector (default = NULL)

mu
The tuning parameter in construction of the projection direction (default = NULL)

step
Number of steps (< maxiter) to obtain the smallest mu that gives convergence of the optimization problem for constructing the projection direction (default = NULL)

resol
Resolution or the factor by which mu is increased/decreased to obtain the smallest mu that gives convergence of the optimization problem of constructing the projection direction (default = 1.5)

maxiter
Maximum number of steps along which mu is increased/decreased to obtain the smallest mu that gives convergence of the optimization problem for constructing the projection direction (default = 10)

Value

prop.est

The bias corrected estimator of the linear functional

se

Standard error of the bias-corrected estimator

proj

The projection direction, of length \(p\)

plug.in

Plug-in LASSO estimator of the linear functional

References

Guo Z, Rakshit P, Herman DS, Chen J (2019). “Inference for Case Probability in High-dimensional Logistic Regression.” Unknown.

Examples

A1gen <- function(rho,p){
A1=matrix(0,p,p)
for(i in 1:p){
  for(j in 1:p){
    A1[i,j]<-rho^(abs(i-j))
  }
}
A1
}
n = 100
p = 400
mu <- rep(0,p)
rho = 0.5
Cov <- (A1gen(rho,p))/2
Cov2<-matrix(NA,nrow=p,ncol=p)
for(i in 1:p){
  for(j in 1:p){
    Cov2[i,j]<-0.5^(1+abs(i-j))
  }
}
beta <- rep(0,p)
beta[1:10] <- c(1:10)/5
X <- MASS::mvrnorm(n,mu,Cov)
exp_val <- X%*%beta
prob <- exp(exp_val)/(1+exp(exp_val))
y <- rbinom(n,1,prob)
loading <- MASS::mvrnorm(1,mu,Cov2)
LF_logistic(X = X, y = y, loading = loading, intercept = TRUE, weight = rep(1,n))

#> Warning: longer object length is not a multiple of shorter object length

#> [1] 4
#> [1] "step is 4"

#> $prop.est
#> [1] -0.9947318
#> 
#> $se
#> [1] 3.334083
#> 
#> $proj
#>   [1]  0.9841238947  0.0795318906  0.8356145557 -3.8013726921  0.8512964508
#>   [6]  0.0382601531 -0.3152590911 -0.1024772124 -0.5704919185 -0.6543544076
#>  [11] -1.0315112820  0.8826596319  0.0580875817 -0.0533063610  0.5005124537
#>  [16]  1.4916448640  0.1888616917 -0.2804808805  0.7888162584  0.6277071173
#>  [21] -0.2205567540  0.4664701190  0.0002170192 -0.8333931354 -1.8180083330
#>  [26] -0.2734689423  0.3510042928  0.8561822671 -0.1268998954 -0.1821489765
#>  [31]  0.1128296563 -0.0600753085  0.2723966206 -0.2716803990  0.5161113163
#>  [36] -2.4338549196 -0.0086823985 -0.0154309634  0.2040659488  0.2396875521
#>  [41] -0.0398082255  0.8557363167  0.5084827396  0.5302973009  0.8517109093
#>  [46]  1.4478616800  0.5094385721  0.8421067385 -0.4590801710 -0.4153425185
#>  [51] -0.9297429108 -1.4573594613 -1.5487979822 -0.3812701603 -0.1781955274
#>  [56] -0.4819493331 -0.3375700604 -1.5759783173 -0.8407449936  1.0399259391
#>  [61]  0.4462443948 -0.1926757169 -0.3522543739 -0.2991703405 -0.0251451827
#>  [66]  0.3593498983  1.1323088535  0.2669715448  0.0450249941 -0.6047171896
#>  [71]  0.0817794222 -0.2830363352  2.6044169933  0.7407691627  0.2392836920
#>  [76] -0.1774062011 -1.1160808082 -0.5568556857  1.1475259431  0.3000780238
#>  [81]  0.2925820235  0.1931711016 -0.0726367021 -0.2330789878 -0.7703215506
#>  [86] -0.4069186825 -1.0465998082 -0.3656893807  0.1869037393 -0.3490278303
#>  [91]  0.3537425875 -0.9657257151  0.3429079548  0.3882884770  1.2990638521
#>  [96] -0.8359338037  1.1691935982  1.1471133783  0.5923239573  0.5135767348
#> [101] -0.1017288055 -0.3845465098  0.0333858659  0.4848822007  0.0741822072
#> [106]  0.4104279507  1.1018276674  0.4174007726  0.6023705578  0.1936245484
#> [111] -0.7529844289 -0.6989392527  1.4476898484 -0.1900776906 -2.1868722441
#> [116] -1.5635570599 -0.3398824461  0.0990323539 -0.0072848245 -0.9305550503
#> [121]  0.3122795634  0.1109413091  0.1457854798  0.6213746382  1.1998293677
#> [126]  0.4825393143  1.4194552041  0.0351961086 -0.0238831449  0.6033728258
#> [131]  0.7437892974  0.7810041433  1.1196478376  0.4518134736  0.0725087810
#> [136] -0.7694682580 -0.2620066189 -0.4645090548 -1.3330435029 -0.8530027188
#> [141] -0.9602881541 -0.2284413878 -0.4816509181 -0.1938469198 -0.1979190392
#> [146] -0.7544304348 -0.1762709050 -0.1331688011  0.1786590552 -0.2519249858
#> [151]  0.2647351031  1.7083979118 -0.0551687948 -1.9007786041  0.6937881511
#> [156]  0.7231727204  1.1846128471  0.2553946512  0.0131224753 -0.2678785619
#> [161] -0.9066282544  1.1801775736 -0.6329609226 -0.1756748964 -0.4393585478
#> [166] -1.1016021970 -1.4822167452 -0.2118307309  2.0526907567 -0.8540367519
#> [171] -0.9088827479  0.0634002555  0.9138442486  0.1786892437  1.0551446664
#> [176]  1.1006661105  1.5717945787  0.7541038103 -0.4897161325  0.2109942295
#> [181]  0.6250578050  0.1536116266  0.6804073957  0.4987352640  0.2032301949
#> [186] -1.1448112253 -1.4882716378 -0.4604755633 -0.1414806283 -0.1606809169
#> [191]  0.7492897693  0.8349942457 -0.2754573540 -0.4537147751 -0.6661982141
#> [196] -0.1432461145 -0.9370696851 -0.5328683972 -0.5961855475 -1.1071298472
#> [201]  0.2421326278 -0.6125409863 -0.1320176419 -0.7807406901 -1.0039771698
#> [206]  0.1426354995  0.1469210941  0.6789292605  0.1020924939  1.7443271855
#> [211]  0.6132776173  1.3724750316  1.4161700847 -0.0256339404 -0.3457871857
#> [216]  0.2444773117  3.4382065092 -0.5168060802  3.2989543451  0.2380747202
#> [221]  0.7584590824  0.9455245787  0.7522239461  1.3717773945  0.4563798935
#> [226] -1.3530954998 -0.4250079686  0.7515172356  0.5600206589  0.9202448129
#> [231] -0.2596027142 -0.8974450528  0.1188587168  0.0160800844 -0.0194159266
#> [236] -0.3802060002  0.5891896221  2.8088178689  2.2231677720  0.2204716957
#> [241] -0.9958311312 -1.4026883501  0.3775444020 -1.4417964704 -0.5439501157
#> [246]  0.7806599072  0.2564534726  0.1867224068 -0.5937841447 -0.2930640199
#> [251] -1.5779020816 -0.1032592159 -0.3527748520 -1.0353370341 -1.1597920944
#> [256]  0.4900093766  0.0257545300  0.0962364739 -0.0557222509 -0.6055586474
#> [261] -0.3802055439 -0.0599381792 -0.4489817970  0.0925381110  0.5071717494
#> [266] -0.4961224660  0.0594056381 -0.3649330133  0.7309759828 -0.3700141122
#> [271] -2.8196541709 -0.8885742246 -0.7854621402 -0.7218002524 -0.5695775966
#> [276] -0.6110208328  0.0135246247  0.4887834363  0.5619732756 -0.1095684566
#> [281]  3.8099263402 -0.5762327588 -1.1648583940 -0.8995503699 -0.4263716115
#> [286] -1.2231493625  0.2787073923 -0.5876168681 -1.1768697170 -1.5159978107
#> [291] -0.8918780818 -0.4001065469 -0.8828387109  0.2481310670 -0.7344632664
#> [296] -1.9538420906 -0.5586984302  2.1311229350  0.8607397446  0.0457137734
#> [301] -0.6393583025  1.1542640024  0.4882691216 -2.2571815532  0.0766925592
#> [306]  0.7130776953  0.1771822789  0.5620871684  0.9372277866 -0.1381673366
#> [311]  0.0839042438 -0.1410358479  0.6049535988  0.2683479663  0.5391995678
#> [316]  0.2755218402  1.6017822632  2.1732860775  0.3882888672  0.7536045051
#> [321] -0.2554961814  0.5394745407 -0.2153597924 -0.5457448243 -1.0592803187
#> [326] -0.0262782834 -4.5203359091 -0.1251747929  0.4204008372 -0.5762750398
#> [331]  0.4174280392  0.4216795678 -0.3589899352  0.7325816163  0.6628746667
#> [336]  0.2754578957  0.7951906040  0.6182006786  0.4725967159 -0.7834404650
#> [341] -0.3857150077  0.1826832204 -0.4318009402 -1.3641076998 -0.4021666764
#> [346]  0.4575670858 -1.4860675085 -0.2149080700  1.4465757522  0.4444271479
#> [351]  1.0228162740  1.1880722070  0.3278150029 -0.0971547391 -1.4978125991
#> [356] -0.9167636929  0.2589525128  0.5986895977  0.0050314189  0.2034437222
#> [361] -0.4264149927 -1.0871807257 -0.8301853867  0.3720574464  1.1243168247
#> [366]  0.0587587361 -1.4286925042 -0.5879015272  1.5699278657 -0.7362715913
#> [371] -0.7639627542 -1.0151172155 -0.1426573538 -0.6175667994  2.1886432464
#> [376]  0.9219404434 -0.1903297427 -1.2956238786 -0.6920022303 -0.4617689339
#> [381]  0.8489787016  0.3446942346  1.5449478577 -1.1088466808  0.2989259491
#> [386]  0.2646103110 -0.7684431594 -1.1129836118 -0.7564084374 -1.4316675174
#> [391] -0.7778195737 -1.0882855224  0.2237016340  0.5712992002  0.1903382066
#> [396] -0.8954780934  1.0658367514  0.3594320466  1.0908967678  0.5724527019
#> [401]  0.2493550335
#> 
#> $step
#> [1] 4
#> 
#> $plug.in
#> [1] 0
#> 
#> $support
#> NULL
#>